% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotateBundles.R
\name{annotateBundles}
\alias{annotateBundles}
\title{annotate a pile of TPMs against a transcriptome bundle (via e.g. EnsDb)}
\usage{
annotateBundles(res, transcriptome, ...)
}
\arguments{
\item{res}{a list of matrices from mergeKallisto}

\item{transcriptome}{character string naming the transcriptome}
}
\value{
a GenomicRanges object with annotations
}
\description{
annotate a pile of TPMs against a transcriptome bundle (via e.g. EnsDb)
}

