% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotateErcc.R
\name{annotateErcc}
\alias{annotateErcc}
\title{annotate spike-in controls for normalization (Life Tech annotations built in)}
\usage{
annotateErcc(res, cols = c("bundle", "name", "egid", "biotype"))
}
\arguments{
\item{res}{a KallistoExperiment resulting from mergeKallisto}

\item{cols}{a character vector listing the metadata columns to pass through}
}
\description{
annotate spike-in controls for normalization (Life Tech annotations built in)
}

