% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotateFeatures.R
\name{annotateFeatures}
\alias{annotateFeatures}
\title{annotate features (genes or transcripts) against (say) EnsemblDb
this is becoming the default dispatcher for almost all annotation}
\usage{
annotateFeatures(kexp, level = c(NA, "gene", "transcript"), ...)
}
\arguments{
\item{kexp}{a kexp}

\item{level}{at what level has the data been summarized? (guess)}
}
\value{
a GRanges, perhaps with annotations for the rows
}
\description{
annotate features (genes or transcripts) against (say) EnsemblDb
this is becoming the default dispatcher for almost all annotation
}

