% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotateGenes.R
\name{annotateGenes}
\alias{annotateGenes}
\title{annotate genes against EnsemblDb (like for annotateEnsembl, but gene-level)}
\usage{
annotateGenes(kexp, transcriptome, ...)
}
\arguments{
\item{kexp}{some sort of *Experiment, perhaps somewhat annotated}

\item{transcriptome}{character string naming the target transcriptome}
}
\value{
the supplied experiment, perhaps further annotated
}
\description{
annotate genes against EnsemblDb (like for annotateEnsembl, but gene-level)
}

