% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotateRepeats.R
\name{annotateRepeats}
\alias{annotateRepeats}
\title{Annotate repeats by class (not finished yet)}
\usage{
annotateRepeats(res, repbase)
}
\arguments{
\item{res}{a SummarizedExperiment from mergeKallisto}

\item{repbase}{a character string naming the repbase version}
}
\description{
Annotate repeats by class (not finished yet)
}

