% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KallistoExperiment-methods.R
\name{as}
\alias{as}
\title{Convert a KallistoExperiment to a SummarizedExperiment without losing data}
\description{
Convert a KallistoExperiment to a SummarizedExperiment without losing data

Convert suitably annotated SummarizedExperiment back to a KallistoExperiment
}

