% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseSpaceFastqUpload.R
\name{baseSpaceFastqUpload}
\alias{baseSpaceFastqUpload}
\title{this script uploads to basespace cloud system,into an open account , and to a project under an authorized user. we assume that the user can import files from SRA, and use fastq-dump to convert sra files into fastq files.  however arkas does support fastq header conversions into BaseSpace standard nomenclature. baseSpaceFastqUpload will check the header of the fastq, if it is an sra header, then arkas will transform the header to illumina standard; further this script will also rename the fastq file itself to illumina standard, and upload it to the basespace cloud.  We assume that the basespace command line interface has been configured, (https://help.basespace.illumina.com/articles/descriptive/basespace-cli/).  this R interface to BaseSpace cloud system is mostly useful for numerous samples targeting the uses of single cell sequencing with samples ranging from upwards 800; thus making fastq uploading difficult; hence the automation of it within Arkas.  FIX ME: we would also like to add the execution of arkas via the basespace CLI to run cloud applications via R.}
\usage{
baseSpaceFastqUpload(headerFormat = c("SRA", "Normal"), fastqPath, fastqFiles)
}
\arguments{
\item{fastqPath, }{character string path to the fastq directory}

\item{fastqFiles}{a vector of sample files to upload into basespace}

\item{headerStyle}{,  character string, where SRA is the header style obtained from importing from SRAdb; the Normal header format is a standard two field record delimited by a space, with each field delimited by several colons.}
}
\description{
this script uploads to basespace cloud system,into an open account , and to a project under an authorized user. we assume that the user can import files from SRA, and use fastq-dump to convert sra files into fastq files.  however arkas does support fastq header conversions into BaseSpace standard nomenclature. baseSpaceFastqUpload will check the header of the fastq, if it is an sra header, then arkas will transform the header to illumina standard; further this script will also rename the fastq file itself to illumina standard, and upload it to the basespace cloud.  We assume that the basespace command line interface has been configured, (https://help.basespace.illumina.com/articles/descriptive/basespace-cli/).  this R interface to BaseSpace cloud system is mostly useful for numerous samples targeting the uses of single cell sequencing with samples ranging from upwards 800; thus making fastq uploading difficult; hence the automation of it within Arkas.  FIX ME: we would also like to add the execution of arkas via the basespace CLI to run cloud applications via R.
}

