% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundledTnxlogFC.R
\name{bundledTnxlogFC}
\alias{bundledTnxlogFC}
\title{Groups bundles of transcripts, discard any that represent pointless tests,
and optionally prune any whose joined bundle IDs tend to choke downstream 
packages for e.g. pathway analyses.  Note that this function may or may not
be optimal for your RNAseq experiment. Please refer to 'Details' for some 
thought exercises about the nature of 'genes'. This function will preserve
the list of transcripts in each gene bundled and show the respective logFC
and limma based statistics in a bedGraph style compatible with IGV viewing;
This allows exploration for visualizing the differential expression for a 
queried gene's transcripts.}
\usage{
bundledTnxlogFC(kexp, bundleID = "gene_id", read.cutoff = 1,
  discardjoined = TRUE, design, queryNumber = 10, ...)
}
\arguments{
\item{kexp}{A KallistoExperiment (or something very much like it)}

\item{bundleID}{The column (in mcols(features(kexp))) of the bundle IDs}

\item{read.cutoff}{Discard transcripts and bundles with < this many counts}

\item{discardjoined}{Discard bundles with IDs "joined" by a ";"?  (TRUE)}

\item{design}{a matrix for running transcript wise analysis to gather limma stats}

\item{queryNumber}{the number from the top ranked genes to select for output}
}
\value{
a matrix of summarized counts per sample bundle
}
\description{
Groups bundles of transcripts, discard any that represent pointless tests,
and optionally prune any whose joined bundle IDs tend to choke downstream 
packages for e.g. pathway analyses.  Note that this function may or may not
be optimal for your RNAseq experiment. Please refer to 'Details' for some 
thought exercises about the nature of 'genes'. This function will preserve
the list of transcripts in each gene bundled and show the respective logFC
and limma based statistics in a bedGraph style compatible with IGV viewing;
This allows exploration for visualizing the differential expression for a 
queried gene's transcripts.
}
\details{
This function sums the estimated counts for each transcript within 
a bundle of transcripts (where "bundle" is a user-defined identifier, often 
but not always a 'gene', sometimes a biotype or a class of repeat elements).
The default approach is to discard all rows where the maximum count is less 
than the specified read.cutoff. Since the default cutoff is 1, this means 
discarding transcripts (and bundles) that were not be detected in any sample.
(Filtering tends to increase statistical power at given false-positive rate)
}
\seealso{
collapseTranscripts
}

