% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTpm.R
\name{calcTpm}
\alias{calcTpm}
\title{given a vector or matrix of est_counts and the same for eff_len, compute TPM}
\usage{
calcTpm(est_counts, eff_len)
}
\arguments{
\item{est_counts}{estimated counts, usually from an HDF5 file}

\item{eff_len}{effective transcipt lengths, usually from an HDF5 file}
}
\value{
either a vector or a matrix of TPMs (matches the args)
}
\description{
given a vector or matrix of est_counts and the same for eff_len, compute TPM
}

