% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/collapseBundles.R
\name{collapseBundles}
\alias{collapseBundles}
\title{collapse bundles of transcripts}
\usage{
collapseBundles(kexp, bundleID = "gene_id", read.cutoff = 1,
  discardjoined = TRUE)
}
\arguments{
\item{kexp}{a KallistoExperiment}

\item{bundleID}{the bundle ID to use}

\item{read.cutoff}{discard transcripts and/or bundles with < this many reads}

\item{discardjoined}{discard bundles with "joined" IDs?  (TRUE)}
}
\value{
a list of summarized counts per sample bundle
}
\description{
collapse bundles of transcripts
}

