% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseByTss.R
\name{collapseByTss}
\alias{collapseByTss}
\title{groundwork for collapsing rows of a RangedSummarizedExperiment assay by TSS}
\usage{
collapseByTss(kexp, groups = c("auto", "sex", "circular"), slop = 200)
}
\arguments{
\item{kexp}{the RSE, in this case a KallistoExperiment}

\item{groups}{groups of chromosomes to keep (default: auto+sex+circular)}

\item{slop}{how wide a range to reduce TSSes over for grouping (200bp)}
}
\value{
the TSS for transcripts on chr1:22/X/Y/M to collapse by TSS,
         or NA for transcripts originating anywhere else (ERCC, repeats, &c)
}
\description{
groundwork for collapsing rows of a RangedSummarizedExperiment assay by TSS
}
