% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseTpm.R
\name{collapseTpm}
\alias{collapseTpm}
\title{Collapse bundles of transcripts, discard any with (default) < 1TPM/bundle, 
and optionally prune any whose joined bundle IDs tend to choke downstream 
packages for e.g. pathway- or network-based enrichment analysis.  Note that 
this function may or may not be optimal for your RNAseq experiment. Please 
refer to 'Details' for some thought exercises about the nature of 'genes'.}
\usage{
collapseTpm(kexp, bundleID = "gene_id", minTPM = 0.01,
  discardjoined = TRUE, tx_biotype = NULL, gene_biotype = NULL,
  biotype_class = NULL, ...)
}
\arguments{
\item{kexp}{A KallistoExperiment (or something very much like it)}

\item{bundleID}{The column (in mcols(features(kexp))) of the bundle IDs}

\item{minTPM}{Discard transcripts/bundles with < this many TPMs (0.01)}

\item{discardjoined}{Discard bundles with IDs "joined" by a ";"? (TRUE)}

\item{tx_biotype}{Restrict to a specific mcols(kexp)$tx_biotype? (NULL)}

\item{gene_biotype}{Restrict to a specific mcols(kexp)$gene_biotype? (NULL)}

\item{biotype_class}{Restrict to a specific mcols(kexp)$biotype_class? (No)}
}
\value{
a matrix of TPMs by bundle for each sample
}
\description{
Collapse bundles of transcripts, discard any with (default) < 1TPM/bundle, 
and optionally prune any whose joined bundle IDs tend to choke downstream 
packages for e.g. pathway- or network-based enrichment analysis.  Note that 
this function may or may not be optimal for your RNAseq experiment. Please 
refer to 'Details' for some thought exercises about the nature of 'genes'.
}
\details{
This function sums transcripts per million (TPM) of each transcript
within bundle of transcripts ("bundle" being a user-defined identifier, often
but not always a 'gene', sometimes a biotype or a class of repeat elements).

The default approach is to discard all rows where the maximum TPM is less 
than the specified cutoff.  Since the default cutoff is 1TPM, this means 
discarding bundles where the total transcripts per million estimate is < 1.
(Filtering tends to increase statistical power at a given false-positive rate
per Bourgon et al, 2010, \link{http://www.pnas.org/content/107/21/9546.long})
}

