% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseTranscripts.R
\name{collapseTranscripts}
\alias{collapseTranscripts}
\title{Collapse transcripts, discarding any with low reads (pointless to test them)
Filtering usually increases statistical power for given false-positive rate;
see Bourgon et al, 2010, \link{http://www.pnas.org/content/107/21/9546.long}.}
\usage{
collapseTranscripts(kexp, read.cutoff = 1, ...)
}
\arguments{
\item{kexp}{A KallistoExperiment (or something very much like it)}

\item{read.cutoff}{Discard transcripts and/or bundles w/ < this many reads}
}
\value{
a matrix of filtered transcript counts
}
\description{
Collapse transcripts, discarding any with low reads (pointless to test them)
Filtering usually increases statistical power for given false-positive rate;
see Bourgon et al, 2010, \link{http://www.pnas.org/content/107/21/9546.long}.
}
\seealso{
collapseBundles
}

