% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CountsAndFeaturesToKallistoExperiment.R
\name{CountsAndFeaturesToKallistoExperiment}
\alias{CountsAndFeaturesToKallistoExperiment}
\title{verbose, specialized alias for SEtoKE so that users can easily find it
FIXME: automate the process of length extraction with usable transcriptome(s)}
\usage{
CountsAndFeaturesToKallistoExperiment(counts, features, transcriptomes, ...)
}
\arguments{
\item{counts}{matrix of transcript or bundle counts}

\item{features}{GRanges of features with valid lengths}

\item{transcriptomes}{mandatory string or strings naming the transcriptomes}

\item{...}{Other stuff (such as covariates=covs and the like)}
}
\value{
a KallistoExperiment
}
\description{
verbose, specialized alias for SEtoKE so that users can easily find it
FIXME: automate the process of length extraction with usable transcriptome(s)
}
\seealso{
SEtoKE
}

