% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KallistoExperiment-methods.R
\docType{methods}
\name{covariates}
\alias{covariates}
\alias{covariates,KallistoExperiment-method}
\alias{covariates<-}
\alias{covariates<-,KallistoExperiment-method}
\alias{pData,KallistoExperiment-method}
\alias{pData<-,KallistoExperiment,DataFrame-method}
\title{covariates generic}
\usage{
covariates(object)

covariates(object) <- value

\S4method{covariates}{KallistoExperiment}(object)

\S4method{pData}{KallistoExperiment}(object)

\S4method{covariates}{KallistoExperiment}(object) <- value

\S4method{pData}{KallistoExperiment,DataFrame}(object) <- value
}
\arguments{
\item{object}{a feature defined object}

\item{value}{a value to replace features}

\item{object}{a feature defined object}
}
\description{
covariates generic

covariates setter/replacement generic
}

