% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumpToBigWig.R
\name{dumpToBigWig}
\alias{dumpToBigWig}
\title{Dump the columns of a kexp to bigWig files.  This requires additional 
machinery beyond the default TxDbLite indexing and as such is not yet 
exported.  We are hopeful that this along with QC metrics can be exported
in a forthcoming release.}
\usage{
dumpToBigWig(kexp, annotations)
}
\arguments{
\item{kexp}{A KallistoExperiment (or something very much like it)}

\item{annotations}{Shared-exon annotations and coordinates for transcripts}
}
\value{
a list of bigWig filenames produced
}
\description{
Dump the columns of a kexp to bigWig files.  This requires additional 
machinery beyond the default TxDbLite indexing and as such is not yet 
exported.  We are hopeful that this along with QC metrics can be exported
in a forthcoming release.
}
\details{
This function sums transcripts per million (TPM) of each transcript
within bundles of transcripts and then writes out a wiggle track of TPM over
exons using the information in the annotations argument.
}

