% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KallistoExperiment-methods.R
\docType{methods}
\name{eff_length}
\alias{eff_length}
\alias{eff_length,KallistoExperiment-method}
\title{eff length generic}
\usage{
eff_length(object)

\S4method{eff_length}{KallistoExperiment}(object)
}
\arguments{
\item{object}{a feature defined object}
}
\description{
eff length generic
}

