% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enrichmentAnalysis.R
\name{enrichmentAnalysis}
\alias{enrichmentAnalysis}
\title{farm out an enrichment analysis to the awesome EnrichmentBrowser package.
for KallistoExperiment objects, the design matrix is used to voom() the
per-gene bundled counts (as in geneWiseAnalysis()) and the voomed log-cpm
values are then fed to ebrowser() as the expression matrix.}
\usage{
enrichmentAnalysis(kexp, design, method = c("ggea", "gsea"),
  species = c("hsa", "mmu"), ...)
}
\arguments{
\item{kexp}{A KallistoExperiment or something like it}

\item{design}{A design matrix, used to transform the counts with voom()}

\item{method}{Network-based (ggea) or set-based (gsea) analysis? ("ggea")}
}
\description{
farm out an enrichment analysis to the awesome EnrichmentBrowser package.
for KallistoExperiment objects, the design matrix is used to voom() the
per-gene bundled counts (as in geneWiseAnalysis()) and the voomed log-cpm
values are then fed to ebrowser() as the expression matrix.
}
\seealso{
\pkg{EnrichmentBrowser}

ebrowser

\pkg{limma}

voom
}

