% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extractIndexName.R
\name{extractIndexName}
\alias{extractIndexName}
\title{extract the transcriptome index used for a Kallisto hdf5 file}
\usage{
extractIndexName(callinfo)
}
\arguments{
\item{callinfo}{the Kallisto call string}
}
\value{
the index name
}
\description{
extract the transcriptome index used for a Kallisto hdf5 file
}

