% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KallistoExperiment-methods.R
\docType{methods}
\name{features}
\alias{features}
\alias{features,KallistoExperiment-method}
\alias{features<-}
\alias{features<-,KallistoExperiment-method}
\title{features generic}
\usage{
features(object)

features(object) <- value

\S4method{features}{KallistoExperiment}(object)

\S4method{features}{KallistoExperiment}(object) <- value
}
\description{
features generic

method for features<-
}

