% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchAppSession.R
\name{fetchAppSession}
\alias{fetchAppSession}
\title{fetch app session variables from BaseSpace JSON}
\usage{
fetchAppSession(jsonFile)
}
\arguments{
\item{jsonFile}{character, the name and/or path to the JSON file}
}
\value{
list     the appSession created from that JSON file
}
\description{
fetch app session variables from BaseSpace JSON
}

