% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fetchKallisto.R
\name{fetchKallisto}
\alias{fetchKallisto}
\title{fetch one sample's worth of Kallisto estimates (ignores bootstraps)
if a txome is specified, annotate and collapse TPM by EGID as well as by tx}
\usage{
fetchKallisto(hdf5File, ...)
}
\arguments{
\item{hdf5File}{the file to read}
}
\description{
fetch one sample's worth of Kallisto estimates (ignores bootstraps)
if a txome is specified, annotate and collapse TPM by EGID as well as by tx
}

