% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fetchRunInfo.R
\name{fetchRunInfo}
\alias{fetchRunInfo}
\title{fetches the JSON-encoded run information for a Kallisto run}
\usage{
fetchRunInfo(ri = "run_info.json")
}
\arguments{
\item{ri}{character string specifying the name of the runinfo file}
}
\description{
fetches the JSON-encoded run information for a Kallisto run
}

