% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findDupes.R
\name{findDupes}
\alias{findDupes}
\title{find any duplicate seqnames in FASTA files BEFORE creating an index...
FIXME: remove them, write out a de-duped merged FASTA, and index that.}
\usage{
findDupes(...)
}
\arguments{
\item{...}{the FASTA file names (may be compressed, doesn't matter)}
}
\value{
data.frame of duplicate seqnames and fasta filenames, else NULL
}
\description{
find any duplicate seqnames in FASTA files BEFORE creating an index...
FIXME: remove them, write out a de-duped merged FASTA, and index that.
}

