% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitRepeats.R
\name{fitRepeats}
\alias{fitRepeats}
\title{repeat analysis is used for analysis of repeat regions from an annotated kexp}
\usage{
fitRepeats(kexp, design...)
}
\arguments{
\item{kexp}{kallisto experiment object}

\item{design}{design matrix for modeling linear fit}
}
\value{
a limma list of linear model statistics
}
\description{
repeat analysis is used for analysis of repeat regions from an annotated kexp
}

