% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitTranscripts.R
\name{fitTranscripts}
\alias{fitTranscripts}
\title{encapsulate limma/voom analysis and TMM normalization at transcript level
FIXME: just farm it out to sleuth, e.g. as seen in ?artemisData::withSleuth}
\usage{
fitTranscripts(kexp, design, read.cutoff = 1, ...)
}
\arguments{
\item{kexp}{A KallistoExperiment}

\item{design}{A model matrix}

\item{read.cutoff}{Exclude transcripts where the maximum count is < this}
}
\value{
A list with elements (design, voomed, fit)
}
\description{
encapsulate limma/voom analysis and TMM normalization at transcript level
FIXME: just farm it out to sleuth, e.g. as seen in ?artemisData::withSleuth
}

