% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatResults.R
\name{formatResults}
\alias{formatResults}
\title{downstream results from fitBundles tabulated into a single object 
showing genes and quantified limma fitted values.}
\usage{
formatResults(res)
}
\arguments{
\item{res}{the output from geneWiseAnalysis}
}
\value{
a single merged object with gene names and limma quantified 
         values for differential expression
}
\description{
example:
}
\details{
\code{ library("artemisData") }
\code{ data("NS", package="artemisData") }
\code{ formatResults(geneWiseAnalysis(NS, exptData(NS)$design)) }
}

