% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromTo.R
\name{fromTo}
\alias{fromTo}
\title{wrap mapIds with some sensible defaults (my idea of sensible)}
\usage{
fromTo(ids, from = "ENTREZID", to = "SYMBOL", org = "Homo.sapiens")
}
\arguments{
\item{ids}{the IDs to map}

\item{from}{what kind of IDs are they?}

\item{to}{what kind of IDs do we want back?}

\item{org}{which organismDb (Homo.sapiens, Mus.musculus, ...) to use?}
}
\value{
mappings of "ids" from type "from" to type "to" for "org"
}
\description{
wrap mapIds with some sensible defaults (my idea of sensible)
}

