% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneWiseAnalysis.R
\name{geneWiseAnalysis}
\alias{geneWiseAnalysis}
\title{Downstream analysis of bundle-aggregated transcript abundance estimates.}
\usage{
geneWiseAnalysis(kexp, design, categories = 10, k = 2, p.cutoff = 0.05,
  species = c("Homo.sapiens", "Mus.musculus"))
}
\arguments{
\item{kexp}{a KallistoExperiment or SummarizedExperiment-like object}

\item{design}{a design matrix w/contrast or coefficient to test in col2}

\item{categories}{how many categories for ReactomeDB enrichment plots (10)}

\item{k}{how many gene clusters to construct for comparison (2)}

\item{p.cutoff}{where to set the p-value cutoff for such plots (0.05)}

\item{species}{which species? (Homo.sapiens; FIX: get from transcriptome)}
}
\description{
Downstream analysis of bundle-aggregated transcript abundance estimates.
}

