% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getKallistoVersion.R
\name{getKallistoVersion}
\alias{getKallistoVersion}
\title{get the running version of Kallisto in the default path (if there is one)}
\usage{
getKallistoVersion()
}
\value{
a string (the version of Kallisto that was found)
}
\description{
get the running version of Kallisto in the default path (if there is one)
}

