% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/indexKallisto.R
\name{indexKallisto}
\alias{indexKallisto}
\title{index transcriptome/transcriptomes (w/MD5 digest to avoid duplicating work)}
\usage{
indexKallisto(fastaFiles, fastaPath)
}
\arguments{
\item{fastaFiles}{a character string or vector of source transcriptomes}

\item{fastaPath}{where to find the preceding FASTA files}
}
\description{
index transcriptome/transcriptomes (w/MD5 digest to avoid duplicating work)
}

