% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isRSE.R
\name{isRSE}
\alias{isRSE}
\title{transition from SummarizedExperiment to RangedSummarizedExperiment is a CF
this function just tests to see which flavor of SE/RSE an object might be}
\usage{
isRSE(x)
}
\arguments{
\item{x}{A SummarizedExperiment-like object of some sort}
}
\value{
(boolean) whether it looks like a RangedSummarizedExperiment
}
\description{
transition from SummarizedExperiment to RangedSummarizedExperiment is a CF
this function just tests to see which flavor of SE/RSE an object might be
}

