% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/KallistoExperiment-class.R, R/KallistoExperiment-methods.R, R/SummarizedExperiment-methods.R
\docType{class}
\name{KallistoExperiment-class}
\alias{KallistoExperiment-class}
\alias{counts,KallistoExperiment-method}
\alias{covariates,ANY-method}
\alias{covariates,KallistoExperiment-method}
\alias{covariates<-,ANY-method}
\alias{eff_length,KallistoExperiment-method}
\alias{features,KallistoExperiment-method}
\alias{features,SummarizedExperiment0-method}
\alias{features<-,KallistoExperiment-method}
\alias{features<-,RangedSummarizedExperiment-method}
\alias{features<-,SummarizedExperiment0-method}
\alias{kallistoVersion,KallistoExperiment-method}
\alias{mad,KallistoExperiment-method}
\alias{tpm,KallistoExperiment-method}
\alias{transcriptomes,KallistoExperiment-method}
\alias{transcriptsBy,KallistoExperiment-method}
\title{A SummarizedExperiment subclass that stores multiple Kallisto runs}
\usage{
\S4method{counts}{KallistoExperiment}(object)

\S4method{covariates}{KallistoExperiment}(object)

\S4method{features}{KallistoExperiment}(x)

\S4method{features}{KallistoExperiment}(object) <- value

\S4method{eff_length}{KallistoExperiment}(object)

\S4method{tpm}{KallistoExperiment}(object)

\S4method{kallistoVersion}{KallistoExperiment}(object)

\S4method{transcriptomes}{KallistoExperiment}(object)

\S4method{transcriptsBy}{KallistoExperiment}(x, by = "gene", ...)

\S4method{mad}{KallistoExperiment}(x)

\S4method{covariates}{ANY}(object)

\S4method{covariates}{ANY}(object) <- value

\S4method{features}{SummarizedExperiment0}(x)

\S4method{features}{RangedSummarizedExperiment}(object) <- value

\S4method{features}{SummarizedExperiment0}(object) <- value
}
\arguments{
\item{object:}{A KallistoExperiment from which to retrieve counts}

\item{object:}{A KallistoExperiment from which to retrieve covariates}

\item{object:}{A KallistoExperiment from which features should be obtained}

\item{object:}{A KallistoExperiment from which features should be obtained}

\item{value:}{Some feature annotations, usually a GRanges or GRangesList}

\item{object:}{A KallistoExperiment with effective transcript lengths}

\item{object:}{A KallistoExperiment with estimated counts & effective lengths}

\item{object:}{A KallistoExperiment}

\item{object:}{A KallistoExperiment}

\item{x:}{A KallistoExperiment}

\item{by:}{The gene_name for which to retrieve transcripts}

\item{object:}{A KallistoExperiment}

\item{object:}{something from which to retrieve covariates}

\item{object:}{something to which covariates should be assigned}

\item{value:}{the covariates to assign (usually a data.frame or DataFrame)}

\item{object:}{something from which features should be obtained}

\item{object:}{something to which features should be assigned}

\item{value:}{the features to assign (usually a GRanges or GRangesList)}
}
\value{
A matrix of counts.

a DataFrame

a GRanges or GRangesList of feature annotations

the KallistoExperiment object, with updated feature annotations

a matrix of effective transcript lengths

a matrix of tpms (transcripts per million)

a string: the version of Kallisto used for pseudoalignment

a string: the transcriptomes against which reads were pseudoaligned

a subset of the object with features whose gene_name matches

a matrix of bootstrapped tx MADs (rows == txs, columns == samples)

the object, perhaps with updated covariates

a GRanges or GRangesList of feature annotations

the object, perhaps with updated feature annotations
}
\description{
FIXME: add RUVg-derived ERCC-calibrated normalization factors in metadata
}
\section{Methods (by generic)}{
\itemize{
\item \code{counts}: Retrieve the estimated count matrix from a KallistoExperiment.

\item \code{covariates}: Retrieve the sample covariates from a KallistoExperiment.

\item \code{features}: Retrieve the per-row annotations for a KallistoExperiment.

\item \code{features<-}: Assign per-row annotations to a KallistoExperiment.

\item \code{eff_length}: Retrieve the matrix of effective transcript lengths from a KallistoExperiment

\item \code{tpm}: Obtain tpm estimates as shown in
https://haroldpimentel.wordpress.com/2014/05/08/what-the-fpkm-a-review-rna-seq-expression-units/

\item \code{kallistoVersion}: Retrieve the version of Kallisto used for alignment from a KallistoExperiment

\item \code{transcriptomes}: Retrieve the transcriptomes used for annotation from a KallistoExperiment

\item \code{transcriptsBy}: Fetch transcripts for a gene, or all transcripts bundled by gene.

\item \code{mad}: Fetch the matrix of MADs for estimated counts, if bootstraps were run.

\item \code{covariates}: 

\item \code{covariates<-}: 

\item \code{features}: 

\item \code{features<-}: 

\item \code{features<-}: 
}}
\section{Slots}{

\describe{
\item{\code{transcriptomes}}{Transcriptomes against which reads were pseudoaligned}

\item{\code{kallistoVersion}}{The version of Kallisto used to pseudoalign the reads}
}}

