% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/KallistoExperiment.R
\name{KallistoExperiment}
\alias{KallistoExperiment}
\title{Initializes a KallistoExperiment and performs some rudimentary checks}
\usage{
KallistoExperiment(est_counts = NULL, eff_length = NULL,
  transcriptomes = NULL, covariates = DataFrame(),
  features = GRangesList(), kallistoVersion = "", est_counts_mad = NULL,
  ...)
}
\arguments{
\item{est_counts}{matrix of estimated counts}

\item{eff_length}{matrix of effective transcript lengths}

\item{transcriptomes}{string or vector of strings naming the target txomes}

\item{covariates}{the column metadata for the object}

\item{features}{the row metadata for the object}

\item{kallistoVersion}{the version of Kallisto used to run the experiment}

\item{est_counts_mad}{matrix of count MADs under the best bootstrap model}
}
\description{
Initializes a KallistoExperiment and performs some rudimentary checks
}

