% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KallistoExperiment-methods.R
\docType{methods}
\name{kallistoVersion}
\alias{kallistoVersion}
\alias{kallistoVersion,KallistoExperiment-method}
\title{kallisto version accessor}
\usage{
kallistoVersion(object)

\S4method{kallistoVersion}{KallistoExperiment}(object)
}
\description{
kallisto version accessor
}

