% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mergeKallisto.R
\name{mergeKallisto}
\alias{mergeKallisto}
\title{merge multiple Kallisto results, by default yielding a SummarizedExperiment}
\usage{
mergeKallisto(sampleDirs, outputPath = ".", txomes = c(),
  value = c("SummarizedExperiment", "list"), ...)
}
\arguments{
\item{sampleDirs}{character vector: directory names holding Kallisto results}

\item{outputPath}{character string: base path to the sampleDirs, default "."}

\item{txomes}{character vector: the target transcriptome(s)/repeatome(s)}

\item{value}{character string: return a SummarizedExperiment or a list?}
}
\description{
merge multiple Kallisto results, by default yielding a SummarizedExperiment
}

