% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R
\docType{class}
\name{MultiAssayExperiment-class}
\alias{MultiAssayExperiment-class}
\title{toy class for enforcing coherence between samples and perSampleMetadata}
\description{
toy class for enforcing coherence between samples and perSampleMetadata
}

