% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairFastqFiles.R
\name{pairFastqFiles}
\alias{.findPairings}
\alias{pairFastqFiles}
\title{figure out how a sample's runs can be properly paired (no SE support yet)}
\usage{
pairFastqFiles(path = ".", extension = "\\\\.fastq\\\\.gz",
  readPrefix = "R", lanePrefix = "L", splitChar = "_", ...)

.findPairings(fragments, ...)
}
\arguments{
\item{path}{a character string specifying where the FASTQ files are}

\item{extension}{what is the file extension?  default is ".fastq.gz"}

\item{readPrefix}{read prefix, usually a single character, default "R"}

\item{lanePrefix}{lane prefix, usually a single character, default "L"}

\item{splitChar}{split character or string, usually "_" for Illumina}

\item{...}{other arguments, passed on to helper functions}
}
\value{
a vector of paired FASTQ files for runKallisto

a data.frame of paired FASTQ filenames
}
\description{
figure out how a sample's runs can be properly paired (no SE support yet)
}
\section{Functions}{
\itemize{
\item \code{.findPairings}: helper function to try and weed through various possible extensions
}}

