% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseFastqNames.R
\name{parseFastqNames}
\alias{parseFastqNames}
\title{like it says: parse the chunks of FASTQ file names to get lanes, fwd/rev, &c}
\usage{
parseFastqNames(allFiles, extension = "\\\\.fastq\\\\.gz", readPrefix = "R",
  lanePrefix = "L", splitChar = "_", ...)
}
\arguments{
\item{allFiles}{all the FASTQ filenames for a sample}

\item{extension}{the file extension, default is ".fastq.gz"}

\item{readPrefix}{the read indicator prefix, default is "R" per Illumina}

\item{lanePrefix}{the lane indicator prefix, default is "L" per Illumina}

\item{splitChar}{the character that is used to separate bits of information}
}
\value{
a data.frame with the sample stubs, read directions & filenames
}
\description{
like it says: parse the chunks of FASTQ file names to get lanes, fwd/rev, &c
}

