% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quasiVoom.R
\name{quasiVoom}
\alias{quasiVoom}
\title{do not use this function yet}
\usage{
quasiVoom(counts, design = NULL, lib.size = NULL,
  normalize.method = "none", plot = FALSE, span = 0.5, ...)
}
\description{
do not use this function yet
}

