% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remapSymbols.R
\name{remapSymbols}
\alias{remapSymbols}
\title{update transcript-/gene-level annotations for a KallistoExperiment or matrix}
\usage{
remapSymbols(x, what = c("transcript", "gene"))
}
\arguments{
\item{x}{the matrix or kexp}

\item{what}{"transcript" or "gene" level reannotation}
}
\value{
x, with updated $gene_name or rownames
}
\description{
update transcript-/gene-level annotations for a KallistoExperiment or matrix
}

