% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeatWiseAnalysis.R
\name{repeatWiseAnalysis}
\alias{repeatWiseAnalysis}
\title{Downstream analysis of bundle-aggregated repeat elements}
\usage{
repeatWiseAnalysis(kexp, design = NULL, how = c("cpm", "tpm"),
  p.cutoff = 0.05, fold.cutoff = 1, read.cutoff = 1,
  species = c("Homo.sapiens", "Mus.musculus"), adjustBy = "holm")
}
\arguments{
\item{kexp}{a KallistoExperiment or SummarizedExperiment-like object}

\item{design}{a design matrix with 2nd coefficient as one to test}

\item{how}{whether to collapse by tpm or cpm}

\item{p.cutoff}{where to set the p-value cutoff for plots, etc. (0.05)}

\item{fold.cutoff}{where to set the log2-FC cutoff for plots, etc. (1==2x)}

\item{read.cutoff}{minimum read coverage (estimated) for a gene bundle}

\item{species}{which species? (Homo.sapiens, Mus.musculus are two currently supported}

\item{adjustBy}{character none, BH,BY, or holm for FDR procedures}
}
\value{
a list w/items design, voomed, fit, top, enriched,
                                  Figures, scaledExprs, clusts, species,
                                  features, ... (perhaps)
}
\description{
Downstream analysis of bundle-aggregated repeat elements
}
\details{
If no design matrix is found, the function will look in 
                   exptData(kexp)$design. If that too is empty it fails.
}
