% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runKallisto.R
\name{runKallisto}
\alias{runKallisto}
\title{run kallisto with fastq.gz files on freshly generated or existing index}
\usage{
runKallisto(sampleDir, indexName = NULL, fastaPath = "/data/fasta",
  fastaFiles = NULL, fastqPath = "/data/input/samples",
  outputPath = "/data/output", bootstraps = 100)
}
\arguments{
\item{sampleDir}{character, subdirectory for sample's FASTQ files}

\item{indexName}{character or NULL, optional name of the index}

\item{fastaPath}{character, where FASTA files are ("/data/fasta")}

\item{fastaFiles}{character vector of FASTA transcriptomes, or NULL}

\item{fastqPath}{character, where sampleDir is ("/data/input/samples")}

\item{outputPath}{character, output in outputPath/sampleDir ("/data/output")}

\item{bootstraps}{integer, how many bootstrap replicates to run? (100)}
}
\description{
run kallisto with fastq.gz files on freshly generated or existing index
}

