% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveArtemisPlots.R
\name{saveArtemisPlots}
\alias{saveArtemisPlots}
\title{Saving plots in PDF format in a single output file}
\usage{
saveArtemisPlots(res, outName)
}
\arguments{
\item{res}{An output from limma created by fitBundles}

\item{outName}{a string for saving the file}
}
\value{
returns a pdf in the working directory
}
\description{
Saving plots in PDF format in a single output file
}

