% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showTnxBundles.R
\name{showTnxBundles}
\alias{showTnxBundles}
\title{Groups bundles of transcripts, discard any that represent pointless tests,
and shows the transcript bundles for each gene.}
\usage{
showTnxBundles(kexp, design, species = c("Homo.sapiens", "Mus.musculus"))
}
\arguments{
\item{kexp}{A KallistoExperiment (or something very much like it)}

\item{design}{the design matrix to be used to fit a linear model for Diff Expr}

\item{species}{either human or mouse}
}
\value{
a matrix of summarized counts per sample bundle
}
\description{
Groups bundles of transcripts, discard any that represent pointless tests,
and shows the transcript bundles for each gene.
}
\seealso{
collapseTranscripts
}

