% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseSpaceFastqUpload.R, R/sraFastqHeaderToIlluminaStandard.R
\name{sraFastqHeaderToIlluminaStandard}
\alias{sraFastqHeaderToIlluminaStandard}
\title{this script prepares fastq files downloaded from SRADb R package, which have sra fastq headers, and changes the sra headers to illumina standard headers as a preparatory step for uploading fastq files into the  cloud system,into an open account , and to a project under an authorized user. we assume that the user can import files from SRA, and use fastq-dump to convert sra files into fastq files.  arkas does support fastq header conversions into BaseSpace standard nomenclature. this method will check the header of the fastq, if it is an sra header, then arkas will transform the header to illumina standard; further this script will also rename the fastq file itself to illumina standard, and prepare the fastq file for an upload to the basespace cloud using the script fastqFileUploadToBaseSpace.R}
\usage{
sraFastqHeaderToIlluminaStandard(headerFormat = c("SRA", "Normal"), fastqPath,
  fastqFile, sraOutputDir, fastqReadNumber = 1)

sraFastqHeaderToIlluminaStandard(headerFormat = c("SRA", "Normal"), fastqPath,
  fastqFile, sraOutputDir, fastqReadNumber = 1)
}
\arguments{
\item{fastqPath, }{character string path to the fastq directory}

\item{fastqFile}{a vector of sample files downloaded from SRA to convert to illumina std}

\item{fastqReadNumber}{integer 1, or 2, this is the read number that will be written in the fastq file, for SRAdb defaults to 1, but we give the option to have 1 or 2 here.}

\item{headerStyle}{,  character string, where SRA is the header style obtained from importing from SRAdb; the Normal header format is a standard two field record delimited by a space, with each field delimited by several colons.}

\item{headerStyle}{,  character string, where SRA is the header style obtained from importing from SRAdb; the Normal header format is a standard two field record delimited by a space, with each field delimited by several colons.}

\item{fastqPath, }{character string path to the fastq directory}

\item{fastqFile}{a vector of sample files to upload into basespace}

\item{fastqReadNumber}{integer 1, or 2, this is the read number that will be written in the fastq file, for SRAdb defaults to 1, but we give the option to have 1 or 2 here.}
}
\value{
a integer defining success or failure

a fastq file with illumina naming and header convention
}
\description{
this script prepares fastq files downloaded from SRADb R package, which have sra fastq headers, and changes the sra headers to illumina standard headers as a preparatory step for uploading fastq files into the  cloud system,into an open account , and to a project under an authorized user. we assume that the user can import files from SRA, and use fastq-dump to convert sra files into fastq files.  arkas does support fastq header conversions into BaseSpace standard nomenclature. this method will check the header of the fastq, if it is an sra header, then arkas will transform the header to illumina standard; further this script will also rename the fastq file itself to illumina standard, and prepare the fastq file for an upload to the basespace cloud using the script fastqFileUploadToBaseSpace.R

this script prepares fastq files for uploading into basespace cloud system,into an open account , and to a project under an authorized user. we assume that the user can import files from SRA, and use fastq-dump to convert sra files into fastq files.  arkas does support fastq header conversions into BaseSpace standard nomenclature. this method will check the header of the fastq, if it is an sra header, then arkas will transform the header to illumina standard; further this script will also rename the fastq file itself to illumina standard, and prepare the fastq for an upload to the basespace cloud.
}

