% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqFileUploadToBaseSpace.R
\docType{data}
\name{sraSingleUpload}
\alias{sraSingleUpload}
\title{once the sraFastqHeaderToIlluminaStandard has been ran on SRA imported samples, the basespace fileupload can be ran. this assumes that a user has a basespace account and has a project directory listed correctly.  We assume that the basespace CLI is installed please see (https://help.basespace.illumina.com/articles/descriptive/basespace-cli/).  this R interface to BaseSpace cloud system is mostly useful for numerous samples targeting the uses of single cell sequencing with samples ranging from upwards 800; thus making fastq uploading difficult; hence the automation of it within Arkas.  FIX ME: we would also like to add the execution of arkas via the basespace CLI to run cloud applications via R.}
\format{An object of class \code{character} of length 1.}
\usage{
sraSingleUpload
}
\description{
once the sraFastqHeaderToIlluminaStandard has been ran on SRA imported samples, the basespace fileupload can be ran. this assumes that a user has a basespace account and has a project directory listed correctly.  We assume that the basespace CLI is installed please see (https://help.basespace.illumina.com/articles/descriptive/basespace-cli/).  this R interface to BaseSpace cloud system is mostly useful for numerous samples targeting the uses of single cell sequencing with samples ranging from upwards 800; thus making fastq uploading difficult; hence the automation of it within Arkas.  FIX ME: we would also like to add the execution of arkas via the basespace CLI to run cloud applications via R.
}
\keyword{datasets}

