% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/str2vec.R
\name{str2vec}
\alias{str2vec}
\title{array-ify a string with a possible leading path/URL/whatever}
\usage{
str2vec(x, y = " ")
}
\arguments{
\item{x}{a string}

\item{y}{a split character (default is " ")}
}
\value{
the array produced by splitting basename(x) on y
}
\description{
array-ify a string with a possible leading path/URL/whatever
}

