% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strpop.R
\name{strpop}
\alias{strpop}
\title{handy string splitting function that operates on basename(x)}
\usage{
strpop(x, y = " ", z = NULL)
}
\arguments{
\item{x}{a string}

\item{y}{a split character (" ")}

\item{z}{an element or elements to return (the last one)}
}
\value{
a string, possibly concatenated from multiple elements
}
\description{
handy string splitting function that operates on basename(x)
}

