% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strpush.R
\name{strpush}
\alias{strpush}
\title{The opposite of strpop: push string(s) onto a string.
unlike Perl or C versions, this function does not modify w behind the scenes}
\usage{
strpush(w, x, y = " ", z = NULL)
}
\arguments{
\item{w}{a string}

\item{x}{one or more elements to push onto the string}

\item{y}{optional, split character (default is " ")}

\item{z}{optional, add x after str2vec(w)[z] (default: length(str2vec(w)))}
}
\value{
a string
}
\description{
The opposite of strpop: push string(s) onto a string.
unlike Perl or C versions, this function does not modify w behind the scenes
}

