% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strshift.R
\name{strshift}
\alias{strshift}
\title{grab the first (or z'th) element(s) of x after splitting it on y
unlike Perl or C versions, this function does not modify x behind the scenes}
\usage{
strshift(x, y = " ", z = 1)
}
\arguments{
\item{x}{a string}

\item{y}{a split character (default is " ")}

\item{z}{which element(s) to retrieve (default is 1)}
}
\value{
the element(s) [z] produced by splitting basename(x) on y
}
\description{
grab the first (or z'th) element(s) of x after splitting it on y
unlike Perl or C versions, this function does not modify x behind the scenes
}

