% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strunshift.R
\name{strunshift}
\alias{strunshift}
\title{shove new element(s) x onto a (possibly splittable on y) string x after [z]
unlike Perl or C versions, this function does not modify w behind the scenes}
\usage{
strunshift(w, x, y = " ", z = 0)
}
\arguments{
\item{w}{a string}

\item{x}{new element(s) to shove onto w}

\item{y}{a split character (default is " ")}

\item{z}{after which element is x appended (default is 0, i.e. prepend x)}
}
\value{
the string produced by shoving x onto str2vec(w, y) after [z]
}
\description{
shove new element(s) x onto a (possibly splittable on y) string x after [z]
unlike Perl or C versions, this function does not modify w behind the scenes
}

