% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SummarizedExperimentToKallistoExperiment.R
\name{SummarizedExperimentToKallistoExperiment}
\alias{SummarizedExperimentToKallistoExperiment}
\title{verbose, specialized alias for SEtoKE so that users can easily find it
FIXME: automate the process of length extraction with usable transcriptome(s)}
\usage{
SummarizedExperimentToKallistoExperiment(SE, transcriptomes)
}
\arguments{
\item{SE}{SummarizedExperiment w/fully annotated rows (features)}

\item{transcriptomes}{mandatory string or strings naming the transcriptomes}
}
\value{
a KallistoExperiment
}
\description{
verbose, specialized alias for SEtoKE so that users can easily find it
FIXME: automate the process of length extraction with usable transcriptome(s)
}
\seealso{
SEtoKE
}

