% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/targetImport.R
\name{targetImport}
\alias{.readTargetFiles}
\alias{targetImport}
\title{import RNAseq data from TARGET Level 3 quantification (isoform or gene level)}
\usage{
targetImport(counts = NULL, files = NULL,
  transcriptome = "EnsDb.Hsapiens.v75", level = c("unknown", "transcripts",
  "genes"), ...)

.readTargetFiles(files, path = ".")
}
\arguments{
\item{counts}{matrix of TARGET counts (NULL, but see details)}

\item{files}{alternatively, file names of TARGET level 3 data (NULL)}

\item{transcriptome}{which transcriptome annotation was used? (Ensembl v75)}

\item{level}{at what level was the data summarized? (guess)}

\item{path}{optionally, where are the files (".")}

\item{files}{the files to import (should all be uniformly formatted)}
}
\value{
a KallistoExperiment

a matrix of raw counts, if all goes well
}
\description{
import RNAseq data from TARGET Level 3 quantification (isoform or gene level)
}
\details{
One or the other of counts OR files must be set!
}
\section{Functions}{
\itemize{
\item \code{.readTargetFiles}: 
}}
\seealso{
CountsAndFeaturesToKallistoExperiment
}

