% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KallistoExperiment-methods.R
\docType{methods}
\name{counts,KallistoExperiment-method}
\alias{counts,KallistoExperiment-method}
\alias{counts,tpm}
\alias{tpm}
\alias{tpm,KallistoExperiment-method}
\alias{tpm,counts}
\title{tpm generic}
\usage{
\S4method{counts}{KallistoExperiment}(object)

tpm(object)

\S4method{tpm}{KallistoExperiment}(object)
}
\description{
tpm generic
}

