% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KallistoExperiment-methods.R
\docType{methods}
\name{transcriptomes}
\alias{transcriptomes}
\alias{transcriptomes,KallistoExperiment-method}
\title{transcriptome accessor}
\usage{
transcriptomes(object)

\S4method{transcriptomes}{KallistoExperiment}(object)
}
\description{
transcriptome accessor
}

