% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transcriptWiseAnalysis.R
\name{transcriptWiseAnalysis}
\alias{transcriptWiseAnalysis}
\title{Downstream analysis of raw transcript abundance estimates.}
\usage{
transcriptWiseAnalysis(kexp, design, p.cutoff = 0.05, fold.cutoff = 1,
  species = c("Homo.sapiens", "Mus.musculus"), annotation = NULL)
}
\arguments{
\item{kexp}{a KallistoExperiment or SummarizedExperiment-like object}

\item{design}{a design matrix w/contrast or coefficient to test in col2}

\item{p.cutoff}{where to set the p-value cutoff for plots, etc. (0.05)}

\item{fold.cutoff}{where to set the log2-FC cutoff for plots, etc. (1 == 2x)}

\item{species}{which species? (Homo.sapiens; FIX: get from transcriptome)}

\item{annotation}{functional annotations for individual transcripts (varies)}
}
\description{
Downstream analysis of raw transcript abundance estimates.
}

