% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strpop}
\alias{strpop}
\title{Utility functions for artemis}
\usage{
strpop(x, y = " ", z = NULL)
}
\arguments{
\item{x}{a string}

\item{y}{a split character (" ")}

\item{z}{an element or elements to return (the last one)}
}
\value{
a string, possibly concatenated from multiple elements
}
\description{
Utility functions for artemis
}
\section{Functions}{
\itemize{
\item \code{strpop}: handy string splitting function that operates on basename(x)
}}

