% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotateBundles.R
\name{annotateBundles}
\alias{annotateBundles}
\title{annotate a pile of tpms against a transcriptome bundle (via e.g. EnsDb)}
\usage{
annotateBundles(kexp, transcriptomes, ...)
}
\arguments{
\item{kexp}{a KallistoExperiment (perhaps somewhat annotated)}

\item{transcriptomes}{character vector (perhaps empty) naming transcriptomes}
}
\value{
a KallistoExperiment (perhaps further annotated)
}
\description{
annotate a pile of tpms against a transcriptome bundle (via e.g. EnsDb)
}

