% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotateBundles.R
\name{annotateBundles}
\alias{annotateBundles}
\title{annotate a pile of TPMs against a transcriptome bundle (via e.g. EnsDb)}
\usage{
annotateBundles(res, txome, ...)
}
\arguments{
\item{res}{a SummarizedExperiment from mergeKallisto}

\item{txome}{a character string naming the txome, e.g. "EnsDb.Hsapiens.v79"}
}
\description{
annotate a pile of TPMs against a transcriptome bundle (via e.g. EnsDb)
}

