% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotateEnsembl.R
\name{annotateEnsembl}
\alias{annotateEnsembl}
\title{annotate a pile of TPMs against an Ensembl transcriptome (via EnsDb)}
\usage{
annotateEnsembl(tpm, txome)
}
\arguments{
\item{tpm}{a matrix of transcripts per million estimates}

\item{txome}{a character string naming the txome, e.g. "EnsDb.Hsapiens.v79"}
}
\description{
annotate a pile of TPMs against an Ensembl transcriptome (via EnsDb)
}

