% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotateEnsembl.R
\name{annotateEnsembl}
\alias{annotateEnsembl}
\title{annotate a pile of TPMs against an Ensembl transcriptome (via EnsDb)}
\usage{
annotateEnsembl(kexp, transcriptome, ...)
}
\arguments{
\item{kexp}{a KallistoExperiment}

\item{transcriptome}{a character string naming the transcriptome}
}
\value{
a (possibly further-annotated) KallistoExperiment
}
\description{
annotate a pile of TPMs against an Ensembl transcriptome (via EnsDb)
}

