% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotateErcc.R
\name{annotateErcc}
\alias{annotateErcc}
\title{annotate the ERCC spike-ins found amongst the transcripts in a run
(artemis includes ThermoFisher/Life annotations for these; see ?ERCC.)}
\usage{
annotateErcc(kexp, ...)
}
\arguments{
\item{kexp}{something that smells vaguely like a KallistoExperiment}
}
\value{
the supplied object, perhaps with ERCC spike-ins annotated
}
\description{
annotate the ERCC spike-ins found amongst the transcripts in a run
(artemis includes ThermoFisher/Life annotations for these; see ?ERCC.)
}
\seealso{
ERCC
}

