% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotateErcc.R
\name{annotateErcc}
\alias{annotateErcc}
\title{annotate spike-in controls for normalization (Life Tech annotations built in)}
\usage{
annotateErcc(res)
}
\arguments{
\item{res}{a SummarizedExperiment resulting from mergeKallisto}
}
\description{
annotate spike-in controls for normalization (Life Tech annotations built in)
}

