% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotateErcc.R
\name{annotateErcc}
\alias{annotateErcc}
\title{annotate the ERCC spike-ins found amongst the transcripts in a Kallisto run
(note: artemis includes ThermoFisher/Life annotations for these controls.)}
\usage{
annotateErcc(kexp, ...)
}
\arguments{
\item{kexp}{a KallistoExperiment}
}
\value{
a KallistoExperiment, perhaps with ERCC spike-ins annotated
}
\description{
annotate the ERCC spike-ins found amongst the transcripts in a Kallisto run
(note: artemis includes ThermoFisher/Life annotations for these controls.)
}

