% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotateFeatures.R
\name{annotateFeatures}
\alias{annotateFeatures}
\title{annotate features (genes or transcripts) against EnsemblDb}
\usage{
annotateFeatures(counts, transcriptome = "EnsDb.Hsapiens.v75",
  level = c("unknown", "gene", "transcript"), ...)
}
\arguments{
\item{counts}{a matrix of counts (or anything with rownames, really)}

\item{transcriptome}{a character string naming the transcriptome}

\item{level}{at what level has the data been summarized? (guess)}
}
\value{
a GRanges, perhaps containing annotations for the rows
}
\description{
annotate features (genes or transcripts) against EnsemblDb
}

