% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotateFeatures.R
\name{annotateFeatures}
\alias{annotateFeatures}
\title{annotate features (genes or transcripts) against (say) EnsemblDb
this is becoming the default dispatcher for almost all annotation}
\usage{
annotateFeatures(kexp, level = c(NA, "gene", "transcript"),
  what = c("KallistoExperiment", "GRanges"), ...)
}
\arguments{
\item{kexp}{a kexp}

\item{level}{at what level has the data been summarized? (guess)}

\item{what}{what data structure shall we return (KallistoExperiment)}
}
\value{
a GRanges or a KallistoExperiment, depending on `what`
}
\description{
annotate features (genes or transcripts) against (say) EnsemblDb
this is becoming the default dispatcher for almost all annotation
}

