% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotateGenes.R
\name{annotateGenes}
\alias{annotateGenes}
\title{annotate genes against EnsemblDb (like for annotateEnsembl, but gene-level)}
\usage{
annotateGenes(kexp, transcriptome, ...)
}
\arguments{
\item{kexp}{a KallistoExperiment (or something that behaves like it)}

\item{transcriptome}{a character string naming the transcriptome}
}
\value{
a possibly further-annotated version of kexp
}
\description{
annotate genes against EnsemblDb (like for annotateEnsembl, but gene-level)
}

