% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotateGenes.R
\name{annotateGenes}
\alias{annotateGenes}
\title{annotate genes against EnsemblDb (like for annotateEnsembl, but gene-level)}
\usage{
annotateGenes(kexp = NULL, transcriptome, counts = NULL, ...)
}
\arguments{
\item{kexp}{a KallistoExperiment (or NULL, if annotating counts)}

\item{transcriptome}{a character string naming the transcriptome}

\item{counts}{a matrix of counts (ignored if !is.null(kexp))}
}
\value{
a possibly further-annotated version of kexp
}
\description{
annotate genes against EnsemblDb (like for annotateEnsembl, but gene-level)
}

