% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotateRepeats.R
\name{annotateRepeats}
\alias{annotateRepeats}
\title{annotate the repeats found amongst unique transcripts in a KallistoExperiment}
\usage{
annotateRepeats(kexp, repeats, ...)
}
\arguments{
\item{kexp}{a KallistoExperiment}

\item{repeats}{the names of the repeat elements in the experiment}
}
\value{
a KallistoExperiment, perhaps with repeats annotated
}
\description{
annotate the repeats found amongst unique transcripts in a KallistoExperiment
}

