% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotateRepeats.R
\name{annotateRepeats}
\alias{annotateRepeats}
\title{annotate the repeats found amongst unique transcripts in a KallistoExperiment}
\usage{
annotateRepeats(kexp, repeatome, ...)
}
\arguments{
\item{kexp}{a KallistoExperiment}

\item{repeatome}{the name of the repeatome (presumably annotated in a db)}
}
\value{
a KallistoExperiment, perhaps with repeats annotated
}
\description{
FIXME: actually annotate some repeats
}

