\name{aRtemis-package}
\alias{aRtemis-package}
\alias{aRtemis}
\docType{package}
\title{
  various utility functions for fast, informative RNAseq analysis
}
\description{
  utility functions for Artemis (hunting companion of Kallisto)
}
\details{
  Forthcoming.
}
\author{
  Tim Triche, Jr. <tim.triche@gmail.com>
}
\references{
  Kallisto: \url{ http://arxiv.org/abs/1505.02710 }
  limma/voom: \url{ http://genomebiology.com/2014/15/2/R29 }
  ssizeRNA: \url{ http://bioinformatics.oxfordjournals.org/content/23/6/739 }
  ReactomePA: \url{ http://www.bioconductor.org/packages/ReactomePA } 
}
\keyword{ package }
\examples{
  # Optional simple examples of the most important functions
}
