% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossModels.R
\name{crossModels}
\alias{crossModels}
\title{Often times it is needed to cross compare edgeR results, limma/voom results across various filtering criteria and comparing normalized results and unnormalized results, which is not a good idea.}
\usage{
crossModels(kexp, crossLevel = c("tx_id", "gene_id"), cutoffMax = 3,
  dataType = c("normalized", "unnormalized"), outputDir = ".",
  design = NULL, p.val = 0.05, adjustBy = "BH",
  species = c("Homo.sapiens", "Mus.musculus"))
}
\arguments{
\item{crossLevel}{character option of tx_id, or gene_id which will compare at the transcript or gene level collapse.}

\item{cutoffMax}{integer, this will be the maximum read.cutoff that will compare each read.cutoff up to the max, i.e. from 1<=cutoffMax thresholding.}

\item{dataType}{character either normalized or unnormalized data to compare. if normalized is selected, then ruv is ran to only compare across normalized results.}

\item{outputDir}{a character path to save all the pdfs printed, includes limma volcano plots, heatmaps.}

\item{design}{a matrix specifying contrasts to compare.}

\item{p.val}{numeric pvalue used for linear fitting}

\item{adjustBy}{character for FDR filters}
}
\value{
returns a data frame and several images to plot.
}
\description{
Often times it is needed to cross compare edgeR results, limma/voom results across various filtering criteria and comparing normalized results and unnormalized results, which is not a good idea.
}

