% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossModels.R
\name{crossModels}
\alias{crossModels}
\title{Often times it is needed to cross compare edgeR results, limma/voom results across various filtering criteria and comparing normalized results and unnormalized results, comparing gene-level, and / or transcript level at various filtering criteria reproducibly. This function can run a cross model generation set across various filtering criteria. by default the design is a treatment factorization where the coeficient is the second column with an intercept term.}
\usage{
crossModels(kexp, crossLevel = c("tx_id", "gene_id"), cutoffMax = 3,
  dataType = c("normalized", "unnormalized"), outputDir = ".",
  design = NULL, p.val = 0.05, adjustBy = "BH",
  species = c("Homo.sapiens", "Mus.musculus"), numberSelected = 200)
}
\arguments{
\item{crossLevel}{character option of tx_id, or gene_id which will compare at the transcript or gene level collapse.}

\item{cutoffMax}{integer, this will be the maximum read.cutoff that will compare each read.cutoff up to the max, i.e. from 1<=cutoffMax thresholding.}

\item{dataType}{character either normalized or unnormalized data to compare. if normalized is selected, then ruv is ran to only compare across normalized results.}

\item{outputDir}{a character path to save all the pdfs printed, includes limma volcano plots, heatmaps.}

\item{design}{a matrix with a treatment level contrasts, does not yet support group-means factorization}

\item{p.val}{numeric for linear fitting}

\item{adjustBy}{character either BH,none,BY,holm}

\item{species}{character Homo.sapiens or Mus.musculus}

\item{numberSelected}{integer, this is the number of the highest ranked adj.P.Val genes to print into a heatmap, the max amount is the number of genes returned from an analysis.}
}
\value{
returns several images plotted.
}
\description{
Often times it is needed to cross compare edgeR results, limma/voom results across various filtering criteria and comparing normalized results and unnormalized results, comparing gene-level, and / or transcript level at various filtering criteria reproducibly. This function can run a cross model generation set across various filtering criteria. by default the design is a treatment factorization where the coeficient is the second column with an intercept term.
}

