% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erccAnalysis.R
\name{erccAnalysis}
\alias{erccAnalysis}
\title{QC plots of ERCC spike-in controls (FIXME: automate RUVSeq normalization?)}
\usage{
erccAnalysis(kexp, ...)
}
\arguments{
\item{kexp}{something that behaves like a KallistoExperiment}
}
\description{
QC plots of ERCC spike-in controls (FIXME: automate RUVSeq normalization?)
}

