% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erccAnalysis.R
\name{erccAnalysis}
\alias{erccAnalysis}
\title{QC plots of ERCC spike-in controls (not exported at the moment)}
\usage{
erccAnalysis(kexp, ...)
}
\arguments{
\item{kexp}{something that behaves like a KallistoExperiment}
}
\value{
returns plots and images of ercc analysis
}
\description{
QC plots of ERCC spike-in controls (not exported at the moment)
}

