% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqFileUploadToBaseSpace.R
\name{fastqFileUploadToBaseSpace}
\alias{fastqFileUploadToBaseSpace}
\title{once the sraFastqHeaderToIlluminaStandard has been ran on SRA imported samples, the basespace fileupload can be ran. this assumes that a user has a basespace account and has a project directory listed correctly.  We assume that the basespace CLI is installed please see (https://help.basespace.illumina.com/articles/descriptive/basespace-cli/).  this R interface to BaseSpace cloud system is mostly useful for numerous samples targeting the uses of single cell sequencing with samples ranging from upwards 800; thus making fastq uploading difficult; hence the automation of it within Arkas. Fastq Headers must be in illumina standards. FIX ME: we would also like to add the execution of arkas via the basespace CLI to run cloud applications via R.}
\usage{
fastqFileUploadToBaseSpace(illuminaDirPath = NULL, illuminafastqFile = NULL,
  basespaceProject = NULL, fastqFileSignature = "_001.fastq.gz",
  illuminaDirs = NULL)
}
\arguments{
\item{illuminaDirPath}{a illumina directory path to illumina standard fastq files with illumina header and naming conventions.}

\item{illuminafastqFile}{fastq files with illumina headers and naming conventions, a vector of file names in illumina standard, this is optional parameter for large directories where hte names are numerious, in this case use the file signature for multi uploads}

\item{basespaceProject}{character string of the basespace project name, this must exist on basespace}

\item{illuminaDirs}{a vector of illumina sample directories which contain illumina fastqs}

\item{fastqFileSignaure}{character that is unique to the fastq file directory where upon grep'ing the desired files will get matched.  the default is the illumina standard suffix _001.fastq.gz which should pick out the illumina files in the case where the fastq directory has multiple raw files.}
}
\value{
nothing, a successful indication that the files were uploaded to basespace
}
\description{
once the sraFastqHeaderToIlluminaStandard has been ran on SRA imported samples, the basespace fileupload can be ran. this assumes that a user has a basespace account and has a project directory listed correctly.  We assume that the basespace CLI is installed please see (https://help.basespace.illumina.com/articles/descriptive/basespace-cli/).  this R interface to BaseSpace cloud system is mostly useful for numerous samples targeting the uses of single cell sequencing with samples ranging from upwards 800; thus making fastq uploading difficult; hence the automation of it within Arkas. Fastq Headers must be in illumina standards. FIX ME: we would also like to add the execution of arkas via the basespace CLI to run cloud applications via R.
}
\examples{
fastqFileUploadToBaseSpace(illuminafastqPath,RNA-123456-1-N_S1_L002_R1_001.fastq.gz,basespaceProject)
}

