% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitBundles.R
\name{fitBundles}
\alias{fitBundles}
\title{encapsulate limma/voom analysis for consistency with ebrowser()}
\usage{
fitBundles(kexp, design, ...)
}
\arguments{
\item{kexp}{A KallistoExperiment}

\item{design}{A model matrix}
}
\value{
A list with elements (design, voomed, fit)
}
\description{
encapsulate limma/voom analysis for consistency with ebrowser()
}

