% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formatResults.R
\name{formatResults}
\alias{formatResults}
\title{downstream results from limma tabulated into a single object showing genes
 and quantified limma fitted values.}
\usage{
formatResults(kexp, res, topGenes)
}
\arguments{
\item{kexp}{a KallistoExperiment or SummarizedExperiment-like object}

\item{res}{a list w/items design, voomed, fit, top, enriched,
scaledExprs, clusts, ... (perhaps)}

\item{topGenes}{rownames(res$top) limma object with keys from species library}
}
\value{
a single merged object with gene names and limma quantified                        values for differential expression
}
\description{
downstream results from limma tabulated into a single object showing genes
 and quantified limma fitted values.
}

