% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneWiseAnalysis.R
\name{geneWiseAnalysis}
\alias{geneWiseAnalysis}
\title{Downstream analysis of bundle-aggregated transcript abundance estimates.}
\usage{
geneWiseAnalysis(kexp, design, p.cutoff = 0.05, fold.cutoff = 1,
  read.cutoff = 1, species = c("Homo.sapiens", "Mus.musculus"))
}
\arguments{
\item{kexp}{a KallistoExperiment or SummarizedExperiment-like object}

\item{design}{a design matrix w/contrast or coefficient to test}

\item{p.cutoff}{where to set the p-value cutoff for plots, etc. (0.05)}

\item{fold.cutoff}{where to set the log2-FC cutoff for plots, etc. (1==2x)}

\item{read.cutoff}{minimum read coverage (estimated) for a gene bundle}

\item{species}{which species? (Homo.sapiens; FIX: get from transcriptome)}

\item{topheat}{how many bundles to include in the cluster heatmaps? (100)}
}
\description{
Downstream analysis of bundle-aggregated transcript abundance estimates.
}

