% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneWiseAnalysis.R
\name{geneWiseAnalysis}
\alias{geneWiseAnalysis}
\title{Downstream analysis of bundle-aggregated transcript abundance estimates.}
\usage{
geneWiseAnalysis(kexp, design = NULL, how = c("cpm", "tpm"),
  p.cutoff = 0.05, fold.cutoff = 1, read.cutoff = 1,
  species = c("Homo.sapiens", "Mus.musculus", "Rattus.norvegicus"), ...)
}
\arguments{
\item{kexp}{a KallistoExperiment or SummarizedExperiment-like object}

\item{design}{a design matrix with 2nd coefficient as one to test}

\item{p.cutoff}{where to set the p-value cutoff for plots, etc. (0.05)}

\item{fold.cutoff}{where to set the log2-FC cutoff for plots, etc. (1==2x)}

\item{read.cutoff}{minimum read coverage (estimated) for a gene bundle}

\item{species}{which species? (Homo.sapiens; FIX: get from transcriptome)}

\item{topheat}{how many bundles to include in the cluster heatmaps? (100)}
}
\value{
a list w/items design, voomed, fit, top, enriched,
                                  Figures, scaledExprs, clusts, species,
                                  features, ... (perhaps)
}
\description{
Downstream analysis of bundle-aggregated transcript abundance estimates.
}
\details{
If no design matrix is found, the function will look in 
                   exptData(kexp)$design. If that too is empty it will fail.
                   There seems to be a bug in rendering Reactome plots, so 
                   it may be necessary to do so manually:  
\code{res <- geneWiseAnalysis(kexp, design, ...)} 
                   followed by 
\code{barplot(res$enriched, showCategory=10)}
                   and 
\code{plot(res$clusts)}
}

