% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneWiseAnalysis.R
\name{geneWiseAnalysis}
\alias{geneWiseAnalysis}
\title{limma-trans for gene-wise DE (like voom but with less bullshit & more auto)
This should make use of Harold's bootstrapping and structSSI to control FDR}
\usage{
geneWiseAnalysis(res, categories = 10, k = 2, p.cutoff = 0.05)
}
\arguments{
\item{res}{a SummarizedExperiment from mergeKallisto}

\item{categories}{how many categories for ReactomeDB enrichment plots}

\item{k}{how many gene clusters to construct for comparison}

\item{p.cutoff}{where to set the p-value cutoff for such plots}
}
\description{
limma-trans for gene-wise DE (like voom but with less bullshit & more auto)
This should make use of Harold's bootstrapping and structSSI to control FDR
}

