% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneWiseAnalysis.R
\name{geneWiseAnalysis}
\alias{geneWiseAnalysis}
\title{Pathway analysis and other similar downstream functions}
\usage{
geneWiseAnalysis(res, design, categories = 10, k = 2, p.cutoff = 0.05,
  ERCC = TRUE, species = c("Homo.sapiens", "Mus.musculus"))
}
\arguments{
\item{res}{a KallistoExperiment from mergeKallisto}

\item{design}{a design matrix with the contrast or coefficient to test in col2}

\item{categories}{how many categories for ReactomeDB enrichment plots}

\item{k}{how many gene clusters to construct for comparison}

\item{p.cutoff}{where to set the p-value cutoff for such plots}

\item{ERCC}{perform ERCC spike-in analysis as well? (default: TRUE)}

\item{species}{which species? (Homo.sapiens; FIX: get from transcriptome)}
}
\description{
Pathway analysis and other similar downstream functions
}

