% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/icgcImport.R
\name{icgcImport}
\alias{icgcImport}
\title{import RNAseq data from ICGC (at least the way it comes from their DCC)}
\usage{
icgcImport(counts, transcriptome, level = c("gene", "transcript"), ...)
}
\arguments{
\item{counts}{A matrix of counts}

\item{transcriptome}{what transcriptome these were derived from}

\item{level}{At what level shall features be annotated? (gene)}
}
\value{
a KallistoExperiment attempting to derive effective lengths
}
\description{
import RNAseq data from ICGC (at least the way it comes from their DCC)
}
\seealso{
CountsAndFeaturesToKallistoExperiment
}

