% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexKallisto.R
\name{indexKallisto}
\alias{indexKallisto}
\title{index transcriptome/transcriptomes}
\usage{
indexKallisto(fastaFiles, fastaPath, fastaTxDbLite = TRUE,
  collapse = "_mergedWith_", kmer = 31, ...)
}
\arguments{
\item{fastaFiles}{a character string or vector of FASTA transcriptomes}

\item{fastaPath}{where to find the preceding FASTA files}

\item{fastaTxDbLite}{boolean: should we try to annotate new FASTAs? (yes)}

\item{collapse}{string to name multi-FASTA indices ("_mergedWith_")}

\item{kmer}{integer, integer 3-31 of kmer size,default 31}
}
\description{
index transcriptome/transcriptomes
}

