% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/indexKallisto.R
\name{indexKallisto}
\alias{.checkIndexDigest}
\alias{.getIndexDigest}
\alias{.getIndexDigestFile}
\alias{.getIndexName}
\alias{.getIndexPath}
\alias{.makeIndexDigest}
\alias{indexKallisto}
\title{index transcriptome/transcriptomes (using an MD5 digest to avoid duplicating work)}
\usage{
indexKallisto(fastaFiles, fastaPath)

.getIndexName(fastaFiles)

.getIndexPath(indexName, fastaPath)

.getIndexDigest(indexPath)

.getIndexDigestFile(indexPath)

.makeIndexDigest(indexPath)

.checkIndexDigest(indexPath)
}
\arguments{
\item{fastaFiles}{a character string or vector of source transcriptomes}

\item{fastaPath}{where to find the preceding FASTA files}

\item{indexName}{string: name of the index (standardized or otherwise)}

\item{indexPath}{string: where the index can supposedly be found}

\item{fastaFiles}{character vector: source transcriptomes}

\item{fastaPath}{string: where the index supposedly can be found}

\item{indexPath}{string: where the index can supposedly be found}

\item{indexPath}{string: where the index can supposedly be found}

\item{indexPath}{string: where the index can supposedly be found}
}
\value{
string: the standardized name for an index from these files

string: the expanded path to the supposed index

string: the MD5 digest of the index (or an error if not found)

string: the filename where the MD5 digest was saved

string: the filename where the MD5 digest was saved

boolean: whether the MD5 file matches the actual digest
}
\description{
index transcriptome/transcriptomes (using an MD5 digest to avoid duplicating work)
}
\section{Functions}{
\itemize{
\item \code{.getIndexName}: 

\item \code{.getIndexPath}: 

\item \code{.getIndexDigest}: 

\item \code{.getIndexDigestFile}: 

\item \code{.makeIndexDigest}: 

\item \code{.checkIndexDigest}: 
}}

