% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/indexKallisto.R
\name{indexKallisto}
\alias{indexKallisto}
\title{index a transcriptome or transcriptomes}
\usage{
indexKallisto(fastaFiles, fastaPath = "/data/fasta")
}
\arguments{
\item{fastaFiles}{a character string or vector of source transcriptomes}

\item{fastaPath}{where to find the preceding FASTA files}
}
\description{
index a transcriptome or transcriptomes
}

