% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/KallistoExperiment-class.R, R/KallistoExperiment-methods.R
\docType{class}
\name{KallistoExperiment-class}
\alias{ERCC,KallistoExperiment-method}
\alias{KallistoExperiment-class}
\alias{TPM,KallistoExperiment-method}
\alias{counts,KallistoExperiment-method}
\alias{covariates,KallistoExperiment-method}
\alias{covariates<-,KallistoExperiment,DataFrame-method}
\alias{covariates<-,KallistoExperiment,data.frame-method}
\alias{eff_length,KallistoExperiment-method}
\alias{features,KallistoExperiment-method}
\alias{features<-,KallistoExperiment,GRangesList-method}
\alias{features<-,KallistoExperiment,GenomicRanges-method}
\title{A SummarizedExperiment subclass that stores multiple Kallisto runs}
\usage{
\S4method{counts}{KallistoExperiment}(object)

\S4method{covariates}{KallistoExperiment}(object)

\S4method{covariates}{KallistoExperiment,DataFrame}(object) <- value

\S4method{covariates}{KallistoExperiment,data.frame}(object) <- value

\S4method{features}{KallistoExperiment}(object)

\S4method{features}{KallistoExperiment,GenomicRanges}(object) <- value

\S4method{features}{KallistoExperiment,GRangesList}(object) <- value

\S4method{eff_length}{KallistoExperiment}(object)

\S4method{TPM}{KallistoExperiment}(object)

\S4method{ERCC}{KallistoExperiment}(object)
}
\arguments{
\item{object:}{A KallistoExperiment from which to retrieve counts}

\item{object:}{A KallistoExperiment from which to retrieve covariates}

\item{object:}{A KallistoExperiment to which covariates should be assigned}

\item{value:}{A DataFrame containing the covariates}

\item{object:}{A KallistoExperiment to which covariates should be assigned}

\item{value:}{A data.frame containing the covariates}

\item{object:}{A KallistoExperiment from which features should be obtained}

\item{object:}{A KallistoExperiment from which features should be obtained}

\item{value:}{A GenomicRanges instance containing feature annotations}

\item{object:}{A KallistoExperiment from which features should be obtained}

\item{value:}{A GRangesList instance containing feature annotations}

\item{object:}{A KallistoExperiment with effective transcript lengths}

\item{object:}{A KallistoExperiment with estimated counts & effective lengths}

\item{object:}{A KallistoExperiment with ERCC spike-in control counts}
}
\value{
the KallistoExperiment object, with updated covariates

the KallistoExperiment object, with updated covariates

a GRanges or GRangesList of feature annotations

the KallistoExperiment object, with updated feature annotations

the KallistoExperiment object, with updated feature annotations

a matrix of effective transcript lengths

a matrix of TPMs (transcripts per million)

a subsetted KallistoExperiment with just the ERCC features
}
\description{
FIXME: add RUVg-derived ERCC-calibrated normalization factors in metadata
}
\section{Methods (by generic)}{
\itemize{
\item \code{counts}: 

\item \code{covariates}: 

\item \code{covariates<-}: 

\item \code{covariates<-}: 

\item \code{features}: 

\item \code{features<-}: 

\item \code{features<-}: 

\item \code{eff_length}: 

\item \code{TPM}: 

\item \code{ERCC}: 
}}
\section{Slots}{

\describe{
\item{\code{transcriptomes}}{Transcriptomes against which reads were pseudoaligned}

\item{\code{kallistoVersion}}{The version of Kallisto used to pseudoalign the reads}
}}

