% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/KallistoExperiment-class.R, R/KallistoExperiment-methods.R, R/SummarizedExperiment-methods.R
\docType{class}
\name{KallistoExperiment-class}
\alias{KallistoExperiment-class}
\alias{TPM,KallistoExperiment-method}
\alias{counts,KallistoExperiment-method}
\alias{covariates,ANY-method}
\alias{covariates,KallistoExperiment-method}
\alias{covariates<-,ANY-method}
\alias{eff_length,KallistoExperiment-method}
\alias{features,ANY-method}
\alias{features,KallistoExperiment-method}
\alias{features<-,ANY-method}
\alias{features<-,KallistoExperiment-method}
\alias{kallistoVersion,KallistoExperiment-method}
\alias{transcriptomes,KallistoExperiment-method}
\title{A SummarizedExperiment subclass that stores multiple Kallisto runs}
\usage{
\S4method{counts}{KallistoExperiment}(object)

\S4method{covariates}{KallistoExperiment}(object)

\S4method{features}{KallistoExperiment}(x)

\S4method{features}{KallistoExperiment}(object) <- value

\S4method{eff_length}{KallistoExperiment}(object)

\S4method{TPM}{KallistoExperiment}(object)

\S4method{kallistoVersion}{KallistoExperiment}(object)

\S4method{transcriptomes}{KallistoExperiment}(object)

\S4method{covariates}{ANY}(object)

\S4method{covariates}{ANY}(object) <- value

\S4method{features}{ANY}(x)

\S4method{features}{ANY}(object) <- value
}
\arguments{
\item{object:}{A KallistoExperiment from which to retrieve counts}

\item{object:}{A KallistoExperiment from which to retrieve covariates}

\item{object:}{A KallistoExperiment from which features should be obtained}

\item{object:}{A KallistoExperiment from which features should be obtained}

\item{value:}{Some feature annotations, usually GRanges or GRangesList}

\item{object:}{A KallistoExperiment with effective transcript lengths}

\item{object:}{A KallistoExperiment with estimated counts & effective lengths}

\item{object:}{A KallistoExperiment}

\item{object:}{A KallistoExperiment}

\item{object:}{something from which to retrieve covariates}

\item{object:}{something to which covariates should be assigned}

\item{value:}{the covariates to assign (usually a data.frame or DataFrame)}

\item{object:}{something from which features should be obtained}

\item{object:}{something to which features should be assigned}

\item{value:}{the features to assign (usually a GRanges or GRangesList)}
}
\value{
a GRanges or GRangesList of feature annotations

the KallistoExperiment object, with updated feature annotations

a matrix of effective transcript lengths

a matrix of TPMs (transcripts per million)

a string: the version of Kallisto used

a string: the transcriptomes against which reads were pseudoaligned

the object, perhaps with updated covariates

a GRanges or GRangesList of feature annotations

the object, perhaps with updated feature annotations
}
\description{
FIXME: add RUVg-derived ERCC-calibrated normalization factors in metadata
}
\section{Methods (by generic)}{
\itemize{
\item \code{counts}: 

\item \code{covariates}: 

\item \code{features}: 

\item \code{features<-}: 

\item \code{eff_length}: 

\item \code{TPM}: 

\item \code{kallistoVersion}: 

\item \code{transcriptomes}: 

\item \code{covariates}: 

\item \code{covariates<-}: 

\item \code{features}: 

\item \code{features<-}: 
}}
\section{Slots}{

\describe{
\item{\code{transcriptomes}}{Transcriptomes against which reads were pseudoaligned}

\item{\code{kallistoVersion}}{The version of Kallisto used to pseudoalign the reads}
}}

