% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KallistoExperiment-class.R
\docType{class}
\name{KallistoExperiment-class}
\alias{KallistoExperiment-class}
\title{A SummarizedExperiment subclass that stores multiple Kallisto runs}
\description{
FIXME: add RUVg-derived ERCC-calibrated normalization factors in metadata
}
\section{Slots}{

\describe{
\item{\code{transcriptomes}}{Transcriptomes against which reads were pseudoaligned}

\item{\code{kallistoVersion}}{The version of Kallisto used to pseudoalign the reads}
}}

