% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KallistoExperiment-class.R, R/KallistoExperiment-methods.R, R/SummarizedExperiment-methods.R
\docType{class}
\name{KallistoExperiment-class}
\alias{(}
\alias{),}
\alias{+}
\alias{-}
\alias{<-}
\alias{Kallisto}
\alias{KallistoExperiment}
\alias{KallistoExperiment-class}
\alias{KallistoExperiment-method}
\alias{KallistoExperiment.}
\alias{Obtain}
\alias{Retrieve}
\alias{\code{}
\alias{a}
\alias{alignment}
\alias{annotations}
\alias{as}
\alias{computed}
\alias{counts,KallistoExperiment-method}
\alias{covariates,ANY-method}
\alias{covariates,KallistoExperiment-method}
\alias{covariates<-,ANY-method}
\alias{covariates<-,KallistoExperiment-method}
\alias{eff_length,KallistoExperiment-method}
\alias{effective}
\alias{exp(rate}
\alias{features,KallistoExperiment-method}
\alias{features,SummarizedExperiment0-method}
\alias{features<-,KallistoExperiment-method}
\alias{features<-,RangedSummarizedExperiment-method}
\alias{features<-,SummarizedExperiment0-method}
\alias{for}
\alias{from}
\alias{https://haroldpimentel.wordpress.com/2014/05/08/what-the-fpkm-a-review-rna-seq-expression-units/}
\alias{in}
\alias{kallistoVersion,}
\alias{kallistoVersion,KallistoExperiment-method}
\alias{lengths}
\alias{log(1e6))}
\alias{log(counts(object))}
\alias{log(eff_length(object));}
\alias{log(sum(exp(rate)))}
\alias{mad,KallistoExperiment-method}
\alias{matrix}
\alias{of}
\alias{pData,KallistoExperiment-method}
\alias{pData<-,KallistoExperiment,DataFrame-method}
\alias{per-row}
\alias{precomputed}
\alias{rate}
\alias{shown}
\alias{specifically,}
\alias{the}
\alias{tpm}
\alias{tpm,KallistoExperiment-method}
\alias{transcript}
\alias{transcriptomes,KallistoExperiment-method}
\alias{transcriptsBy,KallistoExperiment-method}
\alias{used}
\alias{version}
\alias{}}
\title{A SummarizedExperiment subclass that stores multiple Kallisto runs}
\usage{
\S4method{counts}{KallistoExperiment}(object)

covariates(object)

covariates(object) <- value

\S4method{covariates}{KallistoExperiment}(object)

\S4method{pData}{KallistoExperiment}(object)

\S4method{covariates}{KallistoExperiment}(object) <- value

\S4method{pData}{KallistoExperiment,DataFrame}(object) <- value

features(object) <- value

\S4method{features}{KallistoExperiment}(x)

\S4method{features}{KallistoExperiment}(object) <- value

eff_length(object)

\S4method{eff_length}{KallistoExperiment}(object)

tpm(object)

\S4method{tpm}{KallistoExperiment}(object)

kallistoVersion(object)

\S4method{kallistoVersion}{KallistoExperiment}(object)

transcriptomes(object)

\S4method{transcriptomes}{KallistoExperiment}(object)

\S4method{transcriptsBy}{KallistoExperiment}(x, by = "gene", ...)

\S4method{mad}{KallistoExperiment}(x)

\S4method{covariates}{ANY}(object)

\S4method{covariates}{ANY}(object) <- value

\S4method{features}{SummarizedExperiment0}(x)

\S4method{features}{RangedSummarizedExperiment}(object) <- value

\S4method{features}{SummarizedExperiment0}(object) <- value
}
\arguments{
\item{object}{for which features are defined}

\item{object:}{something from which to retrieve covariates to Retrieve the estimated count matrix from a KallistoExperiment.}

\item{object:}{something to which features should be assigned}

\item{value:}{the features to assign (usually a GRanges or GRangesList)}

\item{object:}{something to which features should be assigned}

\item{value:}{the features to assign (usually a GRanges or GRangesList)}

\item{object:}{something to which features should be assigned}

\item{value:}{the features to assign (usually a GRanges or GRangesList)}

\item{object:}{something from which to retrieve covariates}

\item{object:}{something to which covariates should be assigned}

\item{value:}{the covariates to assign (usually a data.frame or DataFrame)}

\item{object:}{something from which features should be obtained}

\item{object:}{something to which features should be assigned}

\item{value:}{the features to assign (usually a GRanges or GRangesList)}

\item{object:}{something to which features should be assigned}

\item{value:}{the features to assign (usually a GRanges or GRangesList)}
}
\value{
the object, perhaps with updated covariates

a GRanges or GRangesList of feature annotations

the object, perhaps with updated feature annotations
}
\description{
FIXME: add RUVg-derived ERCC-calibrated normalization factors in metadata

new generics for arkas covariates definition

generics arkas defined covariates setter

set in GenomicFeatures, which we have to import anyways setGeneric("features", function(object) standardGeneric("features"))

eff_length generic

tpm generic

kallistoVersion generic

transcriptomes generic
}
\section{Methods (by generic)}{
\itemize{
\item \code{counts}: Retrieve the estimated count matrix from a KallistoExperiment.

\item \code{pData}: Retrieve the pData matrix from a KallistoExperiment.

\item \code{covariates<-}: 

\item \code{pData<-}: 

\item \code{features<-}: Assign per-row annotations to a KallistoExperiment.

\item \code{transcriptsBy}: Fetch transcripts for a gene, or all transcripts bundled by gene.

\item \code{mad}: Fetch the matrix of MADs for estimated counts, if bootstraps were run.

\item \code{covariates}: 

\item \code{covariates<-}: 

\item \code{features}: 

\item \code{features<-}: 

\item \code{features<-}: 
}}
\section{Slots}{

\describe{
\item{\code{transcriptomes}}{Transcriptomes against which reads were pseudoaligned}

\item{\code{kallistoVersion}}{The version of Kallisto used to pseudoalign the reads}
}}

