% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KallistoExperiment-class.R, R/SummarizedExperiment-methods.R
\docType{class}
\name{KallistoExperiment-class}
\alias{KallistoExperiment-class}
\alias{covariates,ANY-method}
\alias{covariates<-,ANY-method}
\alias{features,SummarizedExperiment0-method}
\alias{features<-,RangedSummarizedExperiment-method}
\alias{features<-,SummarizedExperiment0-method}
\title{A SummarizedExperiment subclass that stores multiple Kallisto runs}
\usage{
\S4method{covariates}{ANY}(object)

\S4method{covariates}{ANY}(object) <- value

\S4method{features}{SummarizedExperiment0}(object)

\S4method{features}{RangedSummarizedExperiment}(object) <- value

\S4method{features}{SummarizedExperiment0}(object) <- value
}
\arguments{
\item{object}{a feature defined object}

\item{names}{an object with names defined.}

\item{object:}{something from which to retrieve covariates}

\item{object:}{something to which covariates should be assigned}

\item{value:}{the covariates to assign (usually a data.frame or DataFrame)}

\item{object:}{something from which features should be obtained}

\item{object:}{something to which features should be assigned}

\item{value:}{the features to assign (usually a GRanges or GRangesList)}

\item{object:}{something to which features should be assigned}

\item{value:}{the features to assign (usually a GRanges or GRangesList)}
}
\value{
the object, perhaps with updated covariates

a GRanges or GRangesList of feature annotations

the object, perhaps with updated feature annotations
}
\description{
A SummarizedExperiment subclass that stores multiple Kallisto runs
}
\section{Methods (by generic)}{
\itemize{
\item \code{covariates}: 

\item \code{covariates<-}: 

\item \code{features}: 

\item \code{features<-}: 

\item \code{features<-}: 
}}
\section{Slots}{

\describe{
\item{\code{transcriptomes}}{Transcriptomes against which reads were pseudoaligned}

\item{\code{kallistoVersion}}{The version of Kallisto used to pseudoalign the reads}
}}

