% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KallistoExperiment-class.R,
%   R/KallistoExperiment-methods.R
\docType{class}
\name{KallistoExperiment-class}
\alias{KallistoExperiment-class}
\alias{counts,KallistoExperiment-method}
\alias{pData,KallistoExperiment-method}
\alias{pData<-,KallistoExperiment,DataFrame-method}
\alias{KallistoExperiment-class}
\alias{eff_length,KallistoExperiment-method}
\alias{eff_length}
\alias{KallistoExperiment-method}
\alias{KallistoExperiment-class}
\alias{tpm,KallistoExperiment-method}
\alias{tpm}
\alias{KallistoExperiment-method}
\alias{KallistoExperiment-class}
\alias{kallistoVersion,KallistoExperiment-method}
\alias{kallistoVersion}
\alias{KallistoExperiment-method}
\alias{KallistoExperiment-class}
\alias{transcriptomes,KallistoExperiment-method}
\alias{transcriptomes}
\alias{KallistoExperiment-method}
\alias{KallistoExperiment-class}
\alias{mabDev,KallistoExperiment-method}
\alias{mabDev}
\alias{KallistoExperiment-method}
\title{A SummarizedExperiment subclass that stores multiple Kallisto runs}
\usage{
\S4method{counts}{KallistoExperiment}(object)

\S4method{pData}{KallistoExperiment}(object)

\S4method{pData}{KallistoExperiment,DataFrame}(object) <- value

eff_length(object)

\S4method{eff_length}{KallistoExperiment}(object)

tpm(object)

\S4method{tpm}{KallistoExperiment}(object)

kallistoVersion(object)

\S4method{kallistoVersion}{KallistoExperiment}(object)

transcriptomes(object)

\S4method{transcriptomes}{KallistoExperiment}(object)

mabDev(object)

\S4method{mabDev}{KallistoExperiment}(object)
}
\arguments{
\item{object}{KallistoExperiment}

\item{value}{this is the replacement value for object}
}
\description{
A SummarizedExperiment subclass that stores multiple Kallisto runs

returns the estimated counts of a quantified experiment

returns the phenotypic data associated with a KallistoExperiment object

a replacement function specific to phenotypic data

finds the effective length of a transcript entry

finds the transcript per million

finds the kallistoVersion used in quantification

finds the transcriptomes pseudo-aligned against

median abs deviation
}
\section{Slots}{

\describe{
\item{\code{transcriptomes}}{Transcriptomes against which reads were pseudoaligned}

\item{\code{kallistoVersion}}{The version of Kallisto used to pseudoalign the reads}
}}

\section{Slots}{
 
 \describe{
   \item{\code{transcriptomes}:}{ character class \code{"character"}, containing data from transcriptomes}
   \item{\code{kallistoVersion}:}{Object of class \code{"character"}, containing data that needs to go in kallistoVersion.}
 }
}

