% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mergeKallisto.R
\name{mergeKallisto}
\alias{mergeKallisto}
\title{merge several samples' results from Kallisto}
\usage{
mergeKallisto(mergePath, txome = "EnsDb.Hsapiens.v79", ...)
}
\arguments{
\item{mergePath}{where the files are}

\item{txome}{a character string specifying the target Ensembl transcriptome}
}
\description{
merge several samples' results from Kallisto
}

