% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mergeKallisto.R
\name{mergeKallisto}
\alias{mergeKallisto}
\title{Merge multiple Kallisto results, yielding a KallistoExperiment.}
\usage{
mergeKallisto(outputDirs = NULL, outputPath = ".", covariates = NULL,
  transcriptomes = c("ERCC", "EnsDb.Hsapiens.v80", "RepBase.v20_05"),
  parallel = TRUE, ...)
}
\arguments{
\item{outputDirs}{character: directories holding Kallisto results (NULL)}

\item{outputPath}{character: base path to the outputDirs (default is .)}

\item{covariates}{data.frame or DataFrame: per-sample covariates (NULL)}

\item{transcriptomes}{vector: target transcriptomes (EnsDb.Hsapiens.v80)}

\item{parallel}{boolean: try to run the merge in parallel? (TRUE)}
}
\description{
If no transcriptomes are specified (or the ones specified are not found),
the resulting KallistoExperiment will have a rowData/rowRanges object
which consists entirely of unannotated transcripts from chromosome "Unknown".
If transcriptomes are specified and annotations for those transcriptomes are
found, the transcripts described in the annotations will be fully annotated
for transcript ID, gene ID, gene name, entrez ID, and transcript biotype,
provided that these fields are supported in the annotation resources.
}
\details{
FIXME: automatically determine which transcriptomes were used
}

