% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mergeKallisto.R
\name{mergeKallisto}
\alias{mergeKallisto}
\title{merge multiple Kallisto results, yielding a KallistoExperiment}
\usage{
mergeKallisto(sampleDirs = NULL, outputPath = ".", covariates = NULL,
  transcriptome = "EnsDb.Hsapiens.v80", parallel = TRUE, ...)
}
\arguments{
\item{sampleDirs}{character: directories holding Kallisto results (NULL)}

\item{outputPath}{character: base path to the sampleDirs (default is .)}

\item{covariates}{data.frame or DataFrame: per-sample covariates (NULL)}

\item{transcriptome}{character: target transcriptome (EnsDb.Hsapiens.v80)}

\item{parallel}{boolean: try to run the merge in parallel? (TRUE)}
}
\description{
FIXME: automatically determine which transcriptomes were used
}

