% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairFastqFiles.R
\name{pairFastqFiles}
\alias{pairFastqFiles}
\title{figure out how a sample's runs can be properly paired (no SE support yet)}
\usage{
pairFastqFiles(path = ".", extension = "_*", readPrefix = "R")
}
\arguments{
\item{path}{a character string specifying where the FASTQ files are}

\item{extension}{what is the file extension?  default is ".fastq.gz"}

\item{readPrefix}{usually pairs are _R1_/_R2_, so this defaults to "R"}
}
\description{
figure out how a sample's runs can be properly paired (no SE support yet)
}

