% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pathwayPlot.R
\name{pathwayPlot}
\alias{pathwayPlot}
\title{use FGNet and gage to plot an entire pathway's worth of contrasts}
\usage{
pathwayPlot(kexp, method = "gage", groups = NULL, ...)
}
\arguments{
\item{kexp}{a KallistoExperiment}

\item{method}{how to link up gene/transcript IDs with functional terms}

\item{groups}{a factor indicating groups along which results can be split}
}
\description{
use FGNet and gage to plot an entire pathway's worth of contrasts
}

