% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pathwayPlot.R
\name{pathwayPlot}
\alias{pathwayPlot}
\title{use pathview to plot an entire pathway's worth of contrasts}
\usage{
pathwayPlot(res, pathway, ...)
}
\arguments{
\item{res}{results from gene-wise or transcript-wise expression analysis}

\item{pathway}{which pathway (or keywords to find the appropriate pathway)}
}
\description{
use pathview to plot an entire pathway's worth of contrasts
}
\seealso{
EnrichmentBrowser

pathview
}

