% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pathwayPlot.R
\name{pathwayPlot}
\alias{.findPathId}
\alias{.findPathPlot}
\alias{.findSpeciesId}
\alias{.getKEGGIDs}
\alias{.getPathsByName}
\alias{pathwayPlot}
\title{Use pathview to plot a pathway or pathways colored by gene-wise contrasts.
The default is to plot the effect-size-signed, -log10(p.value) for each row.}
\usage{
pathwayPlot(pathway, kexp = NULL, results = NULL, design = NULL,
  coef = 2, IDtype = "ENSEMBL", path = ".", addData = NULL,
  how = "signed", species = c("Homo sapiens", "Mus musculus"), ...)

.findSpeciesId(species, synonyms = NULL)

.getKEGGIDs(species)

.findPathId(pathway, species, quiet = FALSE)

.getPathsByName(pathway, species = "hsa", quiet = FALSE)

.findPathPlot(pathway, species, path = ".")
}
\arguments{
\item{pathway}{the pathway token(s)}

\item{species}{the species ID (hsa)}

\item{quiet}{whether to suppress output (FALSE)}

\item{pathway:}{which pathway (or keywords to find the appropriate pathway)}

\item{kexp:}{a KallistoExperiment (can be null, if \code{results} is OK)}

\item{results:}{cached results from running fitBundles to avoid rerunning it}

\item{design:}{a design matrix to compute contrasts within the experiment}

\item{coef:}{which column in the design matrix to extract coefficients (2)}

\item{IDtype:}{type of identifier for gene/transcript annotations (ENSEMBL)}

\item{path:}{the working directory for downloaded/generated files (".")}

\item{species:}{species for pathway extraction (defaults to Homo sapiens)}

\item{addData:}{optional per-annotation DNA methylation or copy number data}

\item{how:}{how to display effects at each node? (signed -log10(p[j]))}

\item{...:}{additional arguments to be passed to pathview}
}
\value{
a list with the human-readable pathway name(s) and the plot file(s)

pathway names
}
\description{
Use pathview to plot a pathway or pathways colored by gene-wise contrasts.
The default is to plot the effect-size-signed, -log10(p.value) for each row.
}
\section{Functions}{
\itemize{
\item \code{.findSpeciesId}: 

\item \code{.getKEGGIDs}: 

\item \code{.findPathId}: 

\item \code{.getPathsByName}: Some code largely modeled upon the name-to-pathway searching tool seen at:
biobeat.wordpress.com/2013/02/21/accessing-kegg-database-from-rbioconductor/

\item \code{.findPathPlot}: 
}}
\seealso{
enrichmentAnalysis

\pkg{EnrichmentBrowser}

\pkg{pathview}
}

