% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPlot.R
\name{pcaPlot}
\alias{pcaPlot}
\title{plots pca of each kexp assay}
\usage{
pcaPlot(ggFrame = ggInput)
}
\arguments{
\item{ggFrame, }{a data frame of desired components to plot}
}
\value{
a pca plot of an cpm, tpm, length, or median abs. deviations
}
\description{
plots pca of each kexp assay
}

