% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPlot.R
\name{pcaPlot}
\alias{pcaPlot}
\title{plots pca of each kexp assay}
\usage{
pcaPlot(kexpAssays, assayInterested = c("cpm", "tpm", "length", "mad"),
  firstComponent = c("first", "second", "third", "fourth", "fifth", "sixth"),
  secondComponent = c("first", "second", "third", "fourth", "fifth", "sixth"))
}
\arguments{
\item{kexpAssays, }{kallisto experiment assay}

\item{assayInterested, }{one assay of the kexp cpm,m.a.d, etc}

\item{firstComponent}{,  the principal component you wish to compare with second}

\item{secondComponent, }{the principal component you wish to compare with first component}
}
\value{
a pca plot of an cpm, tpm, length, or median abs. deviations
}
\description{
plots pca of each kexp assay
}

