% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPlot.R
\name{pcaPlot}
\alias{pcaPlot}
\title{plots pca of each kexp assay}
\usage{
pcaPlot(kexp, assayInterested = c("cpm", "tpm", "length", "mad"))
}
\arguments{
\item{kexp, }{kallisto experiment object}

\item{assayInterested, }{one assay of the kexp cpm,m.a.d, etc}
}
\value{
a pca plot of an cpm, tpm, length, or median abs. deviations
}
\description{
plots pca of each kexp assay
}

