% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rpkmToTpm.R
\name{rpkmToTpm}
\alias{rpkmToTpm}
\title{Converts RPKM or FPKM estimates to TPM.  Per Colin Dewey,}
\usage{
rpkmToTpm(rpkm, ...)
}

\arguments{
\item{rpkm}{a matrix of RPKM estimates}
}
\value{
a matrix of TPM estimates
}
\description{
tpm[i, j] == (rpkm[i, j]/(sum(rpkm[i, ]))) * 1e6
}
\details{
For nonzero rpkm, this becomes

tpm[i, nonzero] == exp( log(rpkm[i, nonzero]) -
                          log(sum(rpkm[i, nonzero])) +
                          log(1e6) )

and of course if RPKM == 0 then TPM == 0 as well.
}\seealso{
http://bioinformatics.oxfordjournals.org/content/26/4/493.full
}

