% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runKallisto.R
\name{runKallisto}
\alias{runKallisto}
\title{run kallisto with fastq.gz files on a freshly generated or existing index}
\usage{
runKallisto(sampleDir, indexName = NULL, fastaPath = ".",
  fastaFiles = NULL, fastqPath = ".", outputPath = ".",
  bootstraps = 100, threads = 1, bias = TRUE, pseudobam = FALSE,
  singleEnd = FALSE, lengthMean = 150, lengthDev = 0.001,
  collapse = "_mergedWith_", extension = ".fastq.gz", ...)
}
\arguments{
\item{sampleDir}{character, subdirectory for sample's FASTQ files}

\item{indexName}{character or NULL, optional name of the index}

\item{fastaPath}{character, where FASTA files are underneath (".")}

\item{fastaFiles}{character vector of FASTA transcriptomes, or NULL}

\item{fastqPath}{character, where sampleDir is located under (".")}

\item{outputPath}{character, output in outputPath/sampleDir (".")}

\item{bootstraps}{integer, how many bootstrap replicates to run? (100)}

\item{threads}{integer, how many threads to use for bootstraps? (4)}

\item{bias}{boolean, perform bias correction? (TRUE)}

\item{pseudobam}{boolean, produce pseudoBAM output? (FALSE)}

\item{singleEnd}{boolean, produce single end quantification, mean and std.dev required}

\item{lengthMean}{integer, length mean used only for single end quantification}

\item{lengthDev}{integer, length std used only for single end quantification}

\item{collapse}{string to name multi-FASTA indices ("_mergedWith_")}

\item{extension}{string,  to pass the extension into pairFastqFiles()}
}
\description{
run kallisto with fastq.gz files on a freshly generated or existing index
}

