% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runKallisto.R
\name{runKallisto}
\alias{runKallisto}
\title{run kallisto with fastq.gz files on a freshly generated or existing index}
\usage{
runKallisto(sampleDir, indexName = NULL, fastaPath = ".",
  fastaFiles = NULL, fastqPath = ".", outputPath = ".",
  bootstraps = 100, threads = 1, bias = TRUE, pseudobam = FALSE, ...)
}
\arguments{
\item{sampleDir}{character, subdirectory for sample's FASTQ files}

\item{indexName}{character or NULL, optional name of the index}

\item{fastaPath}{character, where FASTA files are underneath (".")}

\item{fastaFiles}{character vector of FASTA transcriptomes, or NULL}

\item{fastqPath}{character, where sampleDir is located under (".")}

\item{outputPath}{character, output in outputPath/sampleDir (".")}

\item{bootstraps}{integer, how many bootstrap replicates to run? (100)}

\item{threads}{integer, how many threads to use for bootstraps? (4)}

\item{bias}{boolean, perform bias correction? (TRUE)}

\item{pseudobam}{boolean, produce pseudoBAM output? (FALSE)}
}
\description{
run kallisto with fastq.gz files on a freshly generated or existing index
}

