% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SEtoKE.R
\name{SEtoKE}
\alias{SEtoKE}
\title{convert a properly annotated SummarizedExperiment with 'counts',
or a matrix of counts and a GRanges of annotations for each count,
to a KallistoExperiment (thereby providing tpm estimates on demand).
Note that, without bias correction, these tpm estimates are somewhat janky.
Note also that the code here is based upon Harold Pimentel's code, cf.
https://haroldpimentel.wordpress.com/2014/05/08/what-the-fpkm-a-review-rna-seq-expression-units/}
\usage{
SEtoKE(SE = NULL, counts = NULL, features = NULL, transcriptomes = NULL)
}
\arguments{
\item{SE}{a properly-annotated SummarizedExperiment}

\item{counts}{if SE is not supplied, a matrix of counts}

\item{features}{if SE is not supplied, GRanges with lengths}

\item{transcriptomes}{mandatory string or strings naming the transcriptomes}
}
\value{
a KallistoExperiment
}
\description{
convert a properly annotated SummarizedExperiment with 'counts',
or a matrix of counts and a GRanges of annotations for each count,
to a KallistoExperiment (thereby providing tpm estimates on demand).
Note that, without bias correction, these tpm estimates are somewhat janky.
Note also that the code here is based upon Harold Pimentel's code, cf.
https://haroldpimentel.wordpress.com/2014/05/08/what-the-fpkm-a-review-rna-seq-expression-units/
}
\seealso{
KallistoExperiment
}

