% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SEtoKE.R
\name{SEtoKE}
\alias{SEtoKE}
\title{Convert a properly annotated SummarizedExperiment with 'counts',
OR a matrix of counts and a GRanges of annotations for each count,
to a KallistoExperiment (thereby providing tpm estimates on demand).}
\usage{
SEtoKE(SE = NULL, counts = NULL, features = NULL, covariates = NULL,
  transcriptomes = NULL, fraglen = 200, ...)
}
\arguments{
\item{SE}{a properly-annotated SummarizedExperiment}

\item{counts}{if SE is not supplied, a matrix of counts}

\item{features}{if SE is not supplied, GRanges with lengths}

\item{covariates}{optional data.frame or DataFrame if SE is not present}

\item{transcriptomes}{mandatory string or strings naming the transcriptomes}

\item{fraglen}{optional mean fragment length estimate for PE runs}
}
\value{
a KallistoExperiment
}
\description{
Note that the code here is based upon Harold Pimentel's code, cf.
https://haroldpimentel.wordpress.com/2014/05/08/what-the-fpkm-a-review-rna-seq-expression-units/
}
\details{
FIXME: add bias correction to eff_lengths, or derive from TPM instead?
FIXME: allow for metadata and per-sample metadata (i.e. MultiAssayExperiment)
FIXME: add some sort of KallistoExperiment method to correct bias post-hoc
}
\seealso{
KallistoExperiment
}

