% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/targetImport.R
\name{targetImport}
\alias{targetImport}
\title{import RNAseq data from TARGET (isoform or gene level)}
\usage{
targetImport(counts, ...)
}
\arguments{
\item{counts}{a matrix of TARGET counts (raw, median normalized, RPKM)}

\item{type}{what type of data it is (default: figure out, ENSG vs ENST)}
}
\value{
a KallistoExperiment
}
\description{
import RNAseq data from TARGET (isoform or gene level)
}
\seealso{
CountsAndFeaturesToKallistoExperiment
}

