% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tcgaImport.R
\name{tcgaImport}
\alias{tcgaImport}
\title{import RNAseqV2 data from TCGA (at least the way it comes from FireHose)}
\usage{
tcgaImport(counts, ...)
}
\arguments{
\item{counts}{a matrix of normalized (eep!) RNAseqV2 counts}
}
\value{
a KallistoExperiment attempting to derive effective lengths
}
\description{
import RNAseqV2 data from TCGA (at least the way it comes from FireHose)
}
\seealso{
CountsAndFeaturesToKallistoExperiment
}

