% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptWiseAnalysis.R
\name{transcriptWiseAnalysis}
\alias{transcriptWiseAnalysis}
\title{Analysis of raw transcript abundance estimates.}
\usage{
transcriptWiseAnalysis(kexp, design, p.cutoff = 0.05, fold.cutoff = 1,
  coef = 2, tx_biotype = NULL, gene_biotype = NULL,
  biotype_class = NULL, ...)
}
\arguments{
\item{kexp}{a KallistoExperiment or SummarizedExperiment-like object}

\item{design}{a design matrix w/contrast or coefficient to test in col2}

\item{p.cutoff}{where to set the p-value cutoff for plots, etc. (0.05)}

\item{fold.cutoff}{where to set the log2-FC cutoff for plots, etc. (1 == 2x)}

\item{coef}{which column of the design matrix to test on (2nd)}

\item{tx_biotype}{optionally restrict to one or more tx_biotype classes}

\item{gene_biotype}{optionally restrict to one or more gene_biotype classes}

\item{biotype_class}{optionally restrict to one or more biotype_class ...es}
}
\description{
Analysis of raw transcript abundance estimates.
}

